Refactor: Split cli into a package
Ninety-three test sites substitute eleven names in cli's namespace, and
only 1557 of its 4188 lines are free of them, so extracting the rest
could never have brought the module under the size limit.  The
substitutable collaborators now live in one module, cli/_deps, reached
through the module object, and the tests follow them there.

Bypassing that indirection fails eleven tests, which is the evidence the
substitutions still reach the code they cover.

The indirection was already wrong in one place: _repo_merge bound
GitHubClient directly, so a test that believed it had replaced the
client would have built a real one against GitHub.  Nothing failed,
because a patch on the wrong namespace still succeeds.  The guard added
in the previous commit finds it.

asyncio needs no such treatment.  Patching cli.asyncio.run sets an
attribute on the stdlib module object, which every caller shares, so
those targets only need a dotted path that resolves; they now name the
submodules that call it.

merge, blocked, _handle_org_merge and _handle_repo_merge are decomposed,
and _handle_org_preview_confirmation's seven parameters become six.  The
Typer command signatures are unchanged, since Typer derives the
command-line interface from them.

Findings drop from 21 to 15, all now in merge_manager.
