--- src/dependamerge/cli.py
@@ -1,4188 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: 2025 The Linux Foundation
-
-import asyncio
-import hashlib
-import logging
-import os
-import sys
-from collections import OrderedDict
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Any
-from urllib.parse import urlparse
-
-import requests
-import typer
-import urllib3.exceptions
-from rich.console import Console
-from rich.table import Table
-
-from ._version import __version__
-from .bot_identity import is_automation_author
-from .close_manager import AsyncCloseManager, CloseResult
-from .error_codes import (
-    DependamergeError,
-    ExitCode,
-    convert_git_error,
-    convert_github_api_error,
-    convert_network_error,
-    exit_for_configuration_error,
-    exit_for_github_api_error,
-    exit_for_pr_state_error,
-    exit_with_error,
-    is_github_api_permission_error,
-    is_network_error,
-)
-from .gerrit import (
-    GerritAuthError,
-    GerritChangeComparator,
-    GerritChangeInfo,
-    GerritComparisonResult,
-    GerritRestError,
-    GerritService,
-    GerritSubmitResult,
-    create_gerrit_comparator,
-    create_gerrit_service,
-    create_submit_manager,
-)
-from .git_ops import GitError
-from .github_async import (
-    GitHubAsync,
-    GraphQLError,
-    RateLimitError,
-    SecondaryRateLimitError,
-)
-from .github_async import (
-    PermissionError as GitHubPermissionError,
-)
-from .github_client import GitHubClient
-from .github_service import AUTOMATION_TOOLS
-from .merge_manager import (
-    DEFAULT_MERGE_TIMEOUT,
-    AsyncMergeManager,
-    MergeResult,
-)
-from .models import ComparisonResult, PullRequestInfo
-from .netrc import (
-    GerritCredentials,
-    NetrcParseError,
-    resolve_gerrit_credentials,
-)
-from .pr_comparator import PRComparator
-from .progress_tracker import MergeProgressTracker, ProgressTracker
-from .resolve_conflicts import FixOptions, FixOrchestrator, PRSelection
-from .rule_violations import (
-    RULE_VIOLATION_MARKER,
-    is_rule_violation,
-    required_status_check_names,
-    required_workflow_names,
-    violation_verb,
-)
-from .system_utils import get_default_workers
-from .url_parser import (
-    ParsedGerritTopicUrl,
-    ParsedOrgUrl,
-    ParsedRepoUrl,
-    ParsedUrl,
-    UrlParseError,
-    parse_change_url,
-    parse_gerrit_topic_url,
-    parse_org_url,
-    parse_owner_arg,
-    parse_repo_url,
-)
-
-# Constants
-MAX_RETRIES = 2
-
-# Default owner-wide global wait ceiling (seconds) for `--max-wait`.
-# 15 minutes balances giving dependabot's rebase + CI time to land
-# against not blocking the shell indefinitely.  0 disables waiting
-# (fire-and-forget).
-DEFAULT_MAX_WAIT = 900.0
-
-
-def version_callback(value: bool):
-    """Callback to show version and exit."""
-    if value:
-        console.print(f"🏷️ dependamerge version {__version__}")
-        raise typer.Exit()
-
-
-class CustomTyper(typer.Typer):
-    """Custom Typer class that shows version in help."""
-
-    def __call__(self, *args, **kwargs):
-        # Check if help is being requested
-        if "--help" in sys.argv or "-h" in sys.argv:
-            console.print(f"🏷️ dependamerge version {__version__}")
-        return super().__call__(*args, **kwargs)
-
-
-app = CustomTyper(
-    help="Find blocked PRs in GitHub organizations and automatically merge pull requests"
-)
-
-
-@app.callback()
-def main(
-    version: bool = typer.Option(
-        False,
-        "--version",
-        callback=version_callback,
-        is_eager=True,
-        help="Show version and exit",
-    ),
-):
-    """
-    Dependamerge command line interface.
-    """
-    # The actual handling is done via the version_callback.
-    # This callback exists only to expose --version at the top level.
-
-
-console = Console(markup=False)
-
-
-def _generate_override_sha(
-    pr_info: PullRequestInfo, commit_message_first_line: str
-) -> str:
-    """
-    Generate a SHA hash based on PR author info and commit message.
-
-    Args:
-        pr_info: Pull request information containing author details
-        commit_message_first_line: First line of the commit message to use as salt
-
-    Returns:
-        First 16 hex characters of a SHA256 hash string.
-    """
-    # Create a string combining author info and commit message first line
-    combined_data = f"{pr_info.author}:{commit_message_first_line.strip()}"
-
-    # Generate SHA256 hash
-    sha_hash = hashlib.sha256(combined_data.encode("utf-8")).hexdigest()
-
-    return sha_hash[:16]
-
-
-def _generate_gerrit_override_sha(change: GerritChangeInfo) -> str:
-    """
-    Generate a SHA hash based on Gerrit change owner and subject.
-
-    Args:
-        change: Gerrit change information containing owner and subject
-
-    Returns:
-        SHA256 hash string
-    """
-    combined_data = f"{change.owner.strip()}:{change.subject.strip()}"
-    sha_hash = hashlib.sha256(combined_data.encode("utf-8")).hexdigest()
-    return sha_hash[:16]
-
-
-def _generate_gerrit_continue_sha(change: GerritChangeInfo) -> str:
-    """
-    Generate a SHA hash for continuing after a Gerrit preview.
-
-    Mirrors _generate_continue_sha for the GitHub path: the value is
-    derived from the source change so the confirmation string is unique
-    per batch and cannot be replayed against a different change.
-
-    Args:
-        change: Source Gerrit change information
-
-    Returns:
-        First 16 hex characters of a SHA256 hash string.
-    """
-    combined_data = (
-        f"continue:{change.project}#{change.number}:{change.subject.strip()}"
-    )
-    sha_hash = hashlib.sha256(combined_data.encode("utf-8")).hexdigest()
-    return sha_hash[:16]
-
-
-def _validate_override_sha(
-    provided_sha: str, pr_info: PullRequestInfo, commit_message_first_line: str
-) -> bool:
-    """
-    Validate that the provided SHA matches the expected one for this PR.
-
-    Args:
-        provided_sha: SHA provided by user via --override flag
-        pr_info: Pull request information
-        commit_message_first_line: First line of commit message
-
-    Returns:
-        True if SHA is valid, False otherwise
-    """
-    expected_sha = _generate_override_sha(pr_info, commit_message_first_line)
-    return provided_sha == expected_sha
-
-
-def _generate_continue_sha(
-    pr_info: PullRequestInfo, commit_message_first_line: str
-) -> str:
-    """
-    Generate a SHA hash for continuing after preview evaluation.
-
-    Args:
-        pr_info: Source pull request information
-        commit_message_first_line: First line of the commit message
-
-    Returns:
-        SHA256 hash string for continuation
-    """
-    # Create a string combining source PR info for preview continuation
-    combined_data = f"continue:{pr_info.repository_full_name}#{pr_info.number}:{commit_message_first_line.strip()}"
-
-    # Generate SHA256 hash
-    sha_hash = hashlib.sha256(combined_data.encode("utf-8")).hexdigest()
-
-    return sha_hash[:16]
-
-
-def _format_condensed_similarity(comparison) -> str:
-    """Format similarity comparison result in condensed format."""
-    reasons = comparison.reasons
-
-    # Check if same author is present
-    has_same_author = any("Same automation author" in reason for reason in reasons)
-
-    score_parts = []
-    for reason in reasons:
-        if "Similar titles (score:" in reason:
-            score = reason.split("score: ")[1].replace(")", "")
-            score_parts.append(f"title {score}")
-        elif "Similar PR descriptions (score:" in reason:
-            score = reason.split("score: ")[1].replace(")", "")
-            score_parts.append(f"descriptions {score}")
-        elif "Similar file changes (score:" in reason:
-            score = reason.split("score: ")[1].replace(")", "")
-            score_parts.append(f"changes {score}")
-
-    if has_same_author:
-        author_text = "Same author; "
-    else:
-        author_text = ""
-
-    total_score = f"total score: {comparison.confidence_score:.2f}"
-
-    if score_parts:
-        breakdown = f" [{', '.join(score_parts)}]"
-    else:
-        breakdown = ""
-
-    return f"{author_text}{total_score}{breakdown}"
-
-
-def _display_change_info(
-    change: GerritChangeInfo,
-    title: str = "",
-    console: Console = console,
-    auth_method: str | None = None,
-) -> None:
-    """Display Gerrit change information in a formatted table.
-
-    Args:
-        change: The Gerrit change info to display.
-        title: Optional title for the table.
-        console: Rich console for output.
-        auth_method: Description of authentication method used (e.g., ".netrc file").
-    """
-
-    table = Table(title=title if title else None)
-    table.add_column("Property", style="cyan")
-    table.add_column("Value", style="green")
-
-    # Map Gerrit status to user-friendly description
-    status_map = {
-        "NEW": "Open (awaiting review)",
-        "MERGED": "Merged",
-        "ABANDONED": "Abandoned",
-    }
-    status_display = status_map.get(change.status, change.status)
-
-    # Check if change is submittable (check merge conflicts first!)
-    if change.status == "NEW":
-        if change.mergeable is False:
-            status_display = "Has merge conflicts"
-        elif change.submittable:
-            status_display = "Ready to submit"
-
-    table.add_row("Project", change.project)
-    table.add_row("Change Number", str(change.number))
-    table.add_row("Subject", change.subject)
-    table.add_row("Owner", change.owner)
-    table.add_row("Branch", change.branch)
-    table.add_row("State", change.status)
-    table.add_row("Status", status_display)
-    if change.files_changed:
-        table.add_row("Files Changed", str(len(change.files_changed)))
-    if change.url:
-        table.add_row("URL", change.url)
-    if auth_method:
-        table.add_row("Auth Method", auth_method)
-
-    console.print(table)
-
-
-def _format_gerrit_similarity(comparison: GerritComparisonResult) -> str:
-    """Format Gerrit comparison result in condensed format."""
-    reasons = comparison.reasons
-
-    # Check if same author is present
-    has_same_author = any("Same automation author" in reason for reason in reasons)
-
-    if has_same_author:
-        author_text = "Same author; "
-    else:
-        author_text = ""
-
-    total_score = f"total score: {comparison.confidence_score:.2f}"
-
-    score_parts = []
-    for reason in reasons:
-        if "Similar subjects" in reason and "score:" in reason:
-            score = reason.split("score: ")[1].replace(")", "")
-            score_parts.append(f"subject {score}")
-        elif "Similar files" in reason and "score:" in reason:
-            score = reason.split("score: ")[1].replace(")", "")
-            score_parts.append(f"files {score}")
-
-    if score_parts:
-        breakdown = f" [{', '.join(score_parts)}]"
-    else:
-        breakdown = ""
-
-    return f"{author_text}{total_score}{breakdown}"
-
-
-def _confirm_gerrit_submission(
-    source_change: GerritChangeInfo,
-    console: Console,
-) -> bool:
-    """Prompt for the continue SHA before a real Gerrit submission.
-
-    Mirrors the interactive preview-then-confirm flow of the GitHub
-    path (_handle_preview_confirmation): the user must type a SHA
-    derived from the source change to proceed.
-
-    Returns:
-        True when the user confirmed and the submission should proceed.
-    """
-    continue_sha = _generate_gerrit_continue_sha(source_change)
-    console.print()
-    console.print(f"To proceed with merging enter: {continue_sha}")
-
-    try:
-        if "pytest" in sys.modules or os.getenv("TESTING"):
-            console.print("⚠️ Test mode detected - skipping interactive prompt")
-            return False
-
-        user_input = input(
-            "Enter the string above to continue (or press Enter to cancel): "
-        ).strip()
-
-        if user_input == continue_sha:
-            return True
-        if user_input == "":
-            console.print("❌ Merge cancelled by user.")
-        else:
-            console.print("❌ Invalid input. Merge cancelled.")
-    except KeyboardInterrupt:
-        console.print("\n❌ Merge cancelled by user.")
-    except EOFError:
-        console.print("\n❌ Merge cancelled.")
-    return False
-
-
-@dataclass
-class _MergeContext:
-    """Shared state threaded through the merge sub-routines."""
-
-    # CLI parameters
-    pr_url: str
-    no_confirm: bool
-    similarity_threshold: float
-    merge_method: str
-    token: str | None
-    override: str | None
-    no_fix: bool
-    merge_timeout: float
-    show_progress: bool
-    debug_matching: bool
-    dismiss_copilot: bool
-    force: str
-    verbose: bool
-    no_netrc: bool
-    netrc_file: Path | None
-    netrc_optional: bool
-    github2gerrit_mode: str
-    include_human_prs: bool = False
-    # Optional behaviour flag; defaulted so a context can be built
-    # without every construction site opting in explicitly.
-    fix_semantic_title: bool = True
-    rebase_local: bool = True
-    # Dry-run: perform the full analysis and preview but never merge,
-    # approve, rebase, or close anything.  Because no write occurs, the
-    # write-permission pre-flight check is skipped so the command can run
-    # under a read-only token (e.g. in CI).  Implies preview-only and
-    # suppresses the interactive confirmation prompt.
-    dry_run: bool = False
-    # Owner-wide global wait ceiling (seconds).  Default 900 (15 min);
-    # 0 = fire-and-forget (arm auto-merge, report pending, never block).
-    # Applies to owner/user-wide runs; ignored for single-PR and
-    # single-repository merges.
-    max_wait: float = DEFAULT_MAX_WAIT
-
-    # Derived / mutable state
-    github_client: GitHubClient | None = None
-    owner: str = ""
-    repo_name: str = ""
-    pr_number: int = 0
-    comparator: PRComparator | None = None
-    source_pr: PullRequestInfo | None = None
-    progress_tracker: MergeProgressTracker | None = None
-    all_similar_prs: list[tuple[PullRequestInfo, ComparisonResult]] = field(
-        default_factory=list
-    )
-
-
-def _validate_merge_inputs(
-    submit_gerrit_changes: bool,
-    skip_gerrit_changes: bool,
-    ignore_github2gerrit: bool,
-    force: str,
-    verbose: bool,
-) -> str:
-    """Validate CLI flags and configure logging.
-
-    Returns the effective github2gerrit_mode string.
-
-    Raises:
-        typer.Exit: On mutually exclusive flags or invalid force level.
-    """
-    g2g_flags_set = sum(
-        [submit_gerrit_changes, skip_gerrit_changes, ignore_github2gerrit]
-    )
-    if g2g_flags_set > 1:
-        console.print(
-            "❌ Error: --submit-gerrit-changes, --skip-gerrit-changes, and "
-            "--ignore-github2gerrit are mutually exclusive."
-        )
-        raise typer.Exit(1)
-
-    if skip_gerrit_changes:
-        github2gerrit_mode = "skip"
-    elif ignore_github2gerrit:
-        github2gerrit_mode = "ignore"
-    else:
-        github2gerrit_mode = "submit"
-
-    if verbose:
-        logging.basicConfig(
-            level=logging.WARNING,
-            format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-        )
-        logging.getLogger("dependamerge").setLevel(logging.DEBUG)
-    else:
-        logging.basicConfig(
-            level=logging.WARNING,
-            format="%(levelname)s - %(message)s",
-        )
-
-    valid_force_levels = [
-        "none",
-        "code-owners",
-        "protection-rules",
-        "all",
-    ]
-    if force not in valid_force_levels:
-        console.print(
-            f"Error: Invalid --force level '{force}'. "
-            f"Must be one of: {', '.join(valid_force_levels)}"
-        )
-        raise typer.Exit(1)
-
-    if force == "all":
-        console.print("⚠️ Warning: Using --force=all will bypass most safety checks.")
-        console.print("   This may attempt merges that will fail at GitHub API level.")
-
-    return github2gerrit_mode
-
-
-def _init_github_merge(ctx: _MergeContext) -> None:
-    """Initialise the GitHub client, progress tracker, and comparator.
-
-    Populates *ctx* in-place with the resolved objects.
-    """
-    ctx.github_client = GitHubClient(ctx.token)
-    assert ctx.github_client.token is not None
-    ctx.token = ctx.github_client.token
-    ctx.owner, ctx.repo_name, ctx.pr_number = ctx.github_client.parse_pr_url(ctx.pr_url)
-
-    if ctx.show_progress:
-        # Create the tracker but do NOT start it yet — it will be
-        # started in _scan_and_find_similar() when the long-running
-        # org scan begins.  The early phases (PR fetch, permissions
-        # check) are fast and use plain console output instead.
-        ctx.progress_tracker = MergeProgressTracker(ctx.owner)
-
-    console.print(f"🔍 Examining source pull request in {ctx.owner}...")
-
-    ctx.comparator = PRComparator(ctx.similarity_threshold)
-
-
-def _fetch_and_validate_source_pr(ctx: _MergeContext) -> None:
-    """Fetch the source PR and validate that it is open.
-
-    Populates *ctx.source_pr*.
-    """
-    assert ctx.github_client is not None
-
-    try:
-        ctx.source_pr = ctx.github_client.get_pull_request_info(
-            ctx.owner, ctx.repo_name, ctx.pr_number
-        )
-        if ctx.source_pr.state != "open":
-            if ctx.progress_tracker:
-                ctx.progress_tracker.stop()
-            exit_for_pr_state_error(
-                ctx.pr_number,
-                "closed",
-                details="Pull request has been closed",
-            )
-    except (
-        urllib3.exceptions.NameResolutionError,
-        urllib3.exceptions.MaxRetryError,
-        requests.exceptions.ConnectionError,
-        requests.exceptions.Timeout,
-        requests.exceptions.RequestException,
-    ) as e:
-        if is_network_error(e):
-            exit_with_error(
-                ExitCode.NETWORK_ERROR,
-                details="Failed to fetch PR details from GitHub API",
-                exception=e,
-            )
-        elif is_github_api_permission_error(e):
-            exit_for_github_api_error(details="Failed to fetch PR details", exception=e)
-        else:
-            exit_with_error(
-                ExitCode.GENERAL_ERROR,
-                message="❌ Failed to fetch PR details",
-                details=str(e),
-                exception=e,
-            )
-
-    assert ctx.source_pr is not None
-
-    _display_pr_info(
-        ctx.source_pr,
-        "",
-        ctx.github_client,
-    )
-
-
-def _source_pr_modifies_workflows(ctx: _MergeContext) -> bool:
-    """Return True when the source PR changes GitHub Actions workflow files.
-
-    Merging such a PR through the REST API requires the classic
-    ``workflow`` token scope (or the fine-grained ``Workflows: Read and
-    write`` permission), which is a *separate* gate from plain repository
-    write access.  Detecting this up-front lets the pre-flight check verify
-    the scope instead of failing only at merge time.
-    """
-    source_pr = ctx.source_pr
-    if source_pr is None:
-        return False
-    for change in source_pr.files_changed:
-        path = getattr(change, "filename", "") or ""
-        if path.startswith(".github/workflows/") and path.endswith((".yml", ".yaml")):
-            return True
-    return False
-
-
-def _report_missing_permissions(
-    missing_perms: list[str],
-    perm_results: dict[str, dict[str, Any]],
-) -> None:
-    """Print grouped guidance for missing token permissions and exit.
-
-    Groups failing operations by their guidance tuple so we don't repeat
-    the same URL block once per failed operation.  In the common 401 /
-    expired-token case every operation lands on the same guidance, which
-    used to produce four near-identical six-line blocks; under the grouped
-    renderer it collapses to a single list of operations followed by one
-    guidance block.  The 403 case (per-operation scope guidance) naturally
-    falls out as one group per distinct guidance tuple, so each operation
-    still gets its own scope hint.
-    """
-    console.print("\n❌ Token Permission Check Failed:\n")
-    groups: OrderedDict[
-        tuple[str | None, str | None, str | None],
-        list[tuple[str, str]],
-    ] = OrderedDict()
-    for op in missing_perms:
-        result = perm_results[op]
-        guidance = result.get("guidance") or {}
-        key = (
-            guidance.get("classic"),
-            guidance.get("fine_grained"),
-            guidance.get("fix"),
-        )
-        groups.setdefault(key, []).append((op, result.get("error", "")))
-
-    for key, items in groups.items():
-        classic, fine_grained, fix = key
-        # List the failing operations in this group.
-        for op, _err in items:
-            console.print(f"   • {op}")
-        # Show the shared error message once if every op in the group
-        # reports the same one (the 401 case); otherwise show each
-        # operation's message inline so distinct failures stay
-        # distinguishable.
-        distinct_errors = {err for _, err in items if err}
-        if len(distinct_errors) == 1:
-            console.print(f"\n   {next(iter(distinct_errors))}")
-        elif distinct_errors:
-            console.print()
-            for op, err in items:
-                if err:
-                    console.print(f"   {op}: {err}")
-        # Single guidance block for the whole group.
-        if classic or fine_grained or fix:
-            console.print()
-            if classic:
-                console.print(f"   Classic:       {classic}")
-            if fine_grained:
-                console.print(f"   Fine-grained:  {fine_grained}")
-            if fix:
-                console.print(f"   Fix:           {fix}")
-        console.print()  # blank line between groups
-
-    console.print("💡 Update your token permissions and try again.")
-    raise typer.Exit(code=3)
-
-
-def _check_merge_permissions(ctx: _MergeContext) -> None:
-    """Pre-flight token permission check.
-
-    Exits early when required permissions are missing.
-    """
-    console.print("🔍 Checking token permissions...")
-
-    async def _check() -> dict[str, dict[str, Any]]:
-        async with GitHubAsync(token=ctx.token) as client:
-            operations = ["approve", "merge", "branch_protection"]
-            if not ctx.no_fix:
-                operations.append("update_branch")
-            # Only assert the workflow scope when the source PR actually
-            # touches workflow files.  The check is a no-op (passes) for
-            # fine-grained / app tokens whose scopes cannot be introspected.
-            if _source_pr_modifies_workflows(ctx):
-                operations.append("merge_workflow")
-            return await client.check_token_permissions(
-                operations, ctx.owner, ctx.repo_name
-            )
-
-    try:
-        perm_results = asyncio.run(_check())
-        # Every permission probed by the pre-flight check is now
-        # required by some part of the merge flow:
-        #
-        # * ``approve``, ``merge``, ``update_branch`` — obviously
-        #   needed for the core merge actions.
-        # * ``branch_protection`` — used by the signature-preserving
-        #   local-rebase gate (``requires_commit_signatures``) and
-        #   by Step 5.5's block-reason analysis.  A token without
-        #   it degrades the merge flow in ways that are confusing
-        #   to diagnose at runtime, so treat it as blocking up-front.
-        #
-        # There is consequently no advisory class — a missing
-        # permission is always a hard failure.  If a future
-        # operation is genuinely optional, partition
-        # ``missing_perms`` into blocking and advisory subsets
-        # (and re-introduce the ``⚠️ … (non-blocking)`` display
-        # for the advisory subset) at that point.
-        missing_perms = [
-            op for op, result in perm_results.items() if not result["has_permission"]
-        ]
-        if missing_perms:
-            _report_missing_permissions(missing_perms, perm_results)
-        console.print("✅ Token has required permissions")
-    except GitHubPermissionError as e:
-        console.print(f"\n❌ Permission check failed: {e}")
-        raise typer.Exit(code=3) from e
-    except typer.Exit:
-        # The hard-failure branch above raises typer.Exit(code=3)
-        # to abort the run.  Without re-raising it here it would
-        # be caught by the broad ``except Exception`` below and
-        # silently downgraded to "Continuing anyway...", letting
-        # the merge proceed against a token we already know lacks
-        # the required permissions.
-        raise
-    except Exception as e:
-        console.print(f"⚠️ Could not verify permissions: {e}")
-        console.print("   Continuing anyway...")
-
-
-def _maybe_check_merge_permissions(ctx: _MergeContext) -> None:
-    """Run the write-permission pre-flight unless this is a dry run.
-
-    A dry run performs no writes (no merge, approve, rebase, or branch
-    update), so the merge/approve/update_branch/branch_protection scopes
-    are not required.  Skipping the check lets the full analysis and
-    preview run under a read-only token — which is exactly what the CI
-    dry-run test matrix needs, since the default workflow token (and
-    other low-privilege tokens) cannot satisfy the write scopes.
-    """
-    if ctx.dry_run:
-        console.print(
-            "🧪 Dry run: skipping token permission check (no changes will be made)"
-        )
-        return
-    _check_merge_permissions(ctx)
-
-
-def _validate_automation_author(ctx: _MergeContext) -> None:
-    """Gate a human-authored source PR behind an explicit opt-in.
-
-    Automation-authored sources pass straight through.  A human-authored
-    source needs either ``--include-human-prs`` (the documented opt-in,
-    which also governs which similar PRs are acted on) or a matching
-    ``--override`` SHA, retained so existing invocations keep working.
-
-    With neither, fail fast.  Previously this printed override guidance
-    and exited *successfully*, which is indistinguishable from "nothing
-    to merge" when scripted, and meant ``--include-human-prs`` appeared
-    to do nothing when pointed at a human-authored PR.
-
-    Raises:
-        SystemExit: When the source is human-authored and unauthorized,
-            or when a supplied override SHA does not match.
-    """
-    assert ctx.github_client is not None
-    assert ctx.source_pr is not None
-
-    if ctx.github_client.is_automation_author(ctx.source_pr.author):
-        return
-
-    human_source_notice = (
-        f"👤 Source PR is human-authored (by {ctx.source_pr.author}); "
-        "proceeding because --include-human-prs was supplied."
-    )
-
-    # Deriving the override SHA costs an extra API call per pull request.
-    # Skip it when --include-human-prs alone already authorizes the run
-    # and there is no supplied override to check it against; at
-    # organisation scale that latency and rate-limit pressure adds up.
-    if ctx.include_human_prs and not ctx.override:
-        console.print(human_source_notice)
-        return
-
-    commit_messages = ctx.github_client.get_pull_request_commits(
-        ctx.owner, ctx.repo_name, ctx.pr_number
-    )
-    first_commit_line = commit_messages[0].split("\n")[0] if commit_messages else ""
-    expected_sha = _generate_override_sha(ctx.source_pr, first_commit_line)
-
-    # A supplied override must match whichever gate ultimately authorizes
-    # the run.  A wrong SHA means the operator is looking at a different
-    # PR than they think, and that is worth stopping for even when
-    # --include-human-prs would otherwise have been sufficient.
-    if ctx.override and not _validate_override_sha(
-        ctx.override, ctx.source_pr, first_commit_line
-    ):
-        exit_with_error(
-            ExitCode.VALIDATION_ERROR,
-            message="❌ Invalid override SHA provided",
-            details=(f"Expected SHA for this PR and author: --override {expected_sha}"),
-        )
-
-    if ctx.include_human_prs:
-        console.print(human_source_notice)
-        return
-
-    if ctx.override:
-        console.print(
-            "Override SHA validated. Proceeding with non-automation PR merge."
-        )
-        console.print(
-            "ℹ️ --include-human-prs is the documented way to authorize "
-            "human-authored PRs; --override remains supported.",
-            style="dim",
-        )
-        return
-
-    exit_with_error(
-        ExitCode.VALIDATION_ERROR,
-        message=(
-            f"❌ Source PR is human-authored (by {ctx.source_pr.author}), "
-            "not from a recognized automation tool"
-        ),
-        details=(
-            "dependamerge acts on automation PRs by default.\n"
-            "To include human-authored PRs, run again with: --include-human-prs\n"
-            f"To authorize only this PR instead: --override {expected_sha}\n"
-            f"That SHA derives from the author '{ctx.source_pr.author}' and "
-            f"commit message '{first_commit_line[:50]}...'"
-        ),
-    )
-
-
-def _scan_and_find_similar(ctx: _MergeContext) -> None:
-    """Scan org repositories and populate *ctx.all_similar_prs*."""
-    assert ctx.github_client is not None
-    assert ctx.source_pr is not None
-    assert ctx.comparator is not None
-
-    console.print(f"Checking owner: {ctx.owner}")
-
-    # Start the progress tracker now — this is where the
-    # long-running owner-wide scan begins.
-    if ctx.progress_tracker:
-        ctx.progress_tracker.start()
-
-    # Repository enumeration and counting is handled internally
-    # by GitHubService via a single-pass GraphQL query that
-    # extracts totalCount on the first page and feeds it to the
-    # progress tracker automatically.
-
-    from .github_service import GitHubService
-
-    async def _find_similar():
-        svc = GitHubService(
-            token=ctx.token,
-            progress_tracker=ctx.progress_tracker,
-            debug_matching=ctx.debug_matching,
-        )
-        try:
-            assert ctx.github_client is not None
-            assert ctx.source_pr is not None
-            assert ctx.comparator is not None
-            only_automation = ctx.github_client.is_automation_author(
-                ctx.source_pr.author
-            )
-            return await svc.find_similar_prs(
-                ctx.owner,
-                ctx.source_pr,
-                ctx.comparator,
-                only_automation=only_automation,
-            )
-        finally:
-            await svc.close()
-
-    ctx.all_similar_prs = asyncio.run(_find_similar())
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-        summary = ctx.progress_tracker.get_summary()
-        elapsed_time = summary.get("elapsed_time")
-        total_prs_analyzed = summary.get("total_prs_analyzed")
-        completed_repositories = summary.get("completed_repositories")
-        similar_prs_found = summary.get("similar_prs_found")
-        errors_count = summary.get("errors_count", 0)
-        console.print(f"\n✅ Analysis completed in {elapsed_time}")
-        console.print(
-            f"📊 Analyzed {total_prs_analyzed} PRs across "
-            f"{completed_repositories} repositories"
-        )
-        console.print(f"🔍 Found {similar_prs_found} similar PRs")
-        if errors_count > 0:
-            console.print(f"⚠️ {errors_count} errors encountered during analysis")
-        # The trailing blank line delineates the similar-PR list that
-        # follows.  When no similar PRs were found there is no list to
-        # separate, so the blank line is suppressed (see below).
-        if ctx.all_similar_prs:
-            console.print()
-    else:
-        console.print(f"\n🔍 Found {len(ctx.all_similar_prs)} similar PRs")
-
-    if not ctx.all_similar_prs:
-        # Not a failure: the supplied PR is simply the only one to
-        # merge.  Use a neutral "skip ahead" glyph rather than ❌ so
-        # the output does not read like an error.
-        console.print("⏩ No similar PRs found for this owner")
-
-    for target_pr, comparison in ctx.all_similar_prs:
-        console.print(f"  • {target_pr.repository_full_name} #{target_pr.number}")
-        console.print(f"    {_format_condensed_similarity(comparison)}")
-
-
-def _run_parallel_merge(
-    ctx: _MergeContext,
-    prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]],
-    preview: bool,
-    concurrency: int = 10,
-    leading_blank: bool = True,
-    repo_scoped: bool = False,
-    stripe: bool = False,
-) -> list[MergeResult]:
-    """Execute a parallel merge (preview or real) and return results.
-
-    Args:
-        ctx: Shared merge context.
-        prs_to_merge: PRs to process.
-        preview: If True, run in preview mode without side effects.
-        concurrency: Maximum number of concurrent merge workers.
-            For org-wide merges (PRs spread across repos) the default
-            of 10 is fine.  For repo-scoped merges (all PRs in the
-            same repo) ``AsyncMergeManager`` serialises the actual
-            ``merge_pull_request`` API call via a per-repo dispatch
-            lock, so a value > 1 is safe and lets the worker pool
-            keep processing other PRs while one PR sits in Step 5.5's
-            auto-merge wait loop — see ``_get_merge_dispatch_lock``.
-        leading_blank: When True (default), the ``🚀 Merging ...``
-            banner is preceded by a blank line to separate it from a
-            preceding similar-PR list.  Callers pass False when no
-            list was printed (e.g. no similar PRs were found) so the
-            output stays compact.
-        repo_scoped: When True, each worker refreshes its PR's live
-            merge state just before dispatching the merge, because a
-            sibling merge in the same repository can make a PR ``dirty``
-            / ``behind`` mid-batch.  Enabled for both repo-scoped merges
-            (all PRs in one repository) and owner-wide striped merges
-            (which interleave repositories but serialise PRs within each
-            repository).  See
-            ``AsyncMergeManager._refresh_pr_mergeability``.
-        stripe: When True, the merge is scheduled with the striped
-            scheduler (one serial worker per repository, distinct repos
-            concurrent) so owner-wide batches never stack two merges on
-            the same repository at once.  See
-            ``AsyncMergeManager._run_striped``.
-    """
-
-    async def _do_merge():
-        async with AsyncMergeManager(
-            token=ctx.token,  # pyright: ignore[reportArgumentType]
-            merge_method=ctx.merge_method,
-            max_retries=MAX_RETRIES,
-            concurrency=concurrency,
-            fix_out_of_date=not ctx.no_fix,
-            fix_semantic_title=ctx.fix_semantic_title,
-            merge_timeout=ctx.merge_timeout,
-            progress_tracker=ctx.progress_tracker,
-            preview_mode=preview,
-            dismiss_copilot=ctx.dismiss_copilot,
-            force_level=ctx.force,
-            github2gerrit_mode=ctx.github2gerrit_mode,
-            no_netrc=ctx.no_netrc,
-            netrc_file=ctx.netrc_file,
-            rebase_local=ctx.rebase_local,
-            repo_scoped=repo_scoped,
-            # The owner-wide global wait ceiling applies to striped
-            # (owner/user-wide) runs; single-PR and single-repository
-            # merges keep the uncapped per-PR ``merge_timeout`` behaviour.
-            max_wait=ctx.max_wait if stripe else None,
-        ) as merge_manager:
-            if not preview:
-                prefix = "\n" if leading_blank else ""
-                console.print(
-                    f"{prefix}🚀 Merging {len(prs_to_merge)} pull requests..."
-                )
-            return await merge_manager.merge_prs_parallel(prs_to_merge, stripe=stripe)
-
-    return asyncio.run(_do_merge())
-
-
-def _restart_merge_progress_tracker(ctx: _MergeContext, total_prs: int) -> None:
-    """Stand up a fresh, started progress tracker for the merge phase.
-
-    The org-wide scan (``_scan_and_find_similar``) calls ``stop()`` on
-    the tracker it used, tearing down its Rich ``Live`` and leaving
-    ``ctx.progress_tracker`` in a stopped state.  Reusing that stopped
-    tracker for the real merge means the background wait-status ticker
-    pushes ``update_operation`` calls into a dead ``Live`` — silently
-    dropped by ``ProgressTracker._refresh_display`` — so the user sees
-    no countdown while PRs sit in the Step 5.5 auto-merge wait.
-
-    Mirror the repo-scoped path (``_handle_repo_merge`` /
-    ``_execute_repo_confirmed_merge``) by replacing the tracker with a
-    fresh one dedicated to the merge phase and starting it.  No-op when
-    progress display is disabled (``--no-progress``), where the plain
-    text ticker already provides feedback.
-    """
-    if not ctx.show_progress:
-        return
-    ctx.progress_tracker = MergeProgressTracker(
-        ctx.owner,
-        operation_label="Merging PRs",
-        operation_icon="▶️",
-    )
-    ctx.progress_tracker.set_total_prs(total_prs)
-    ctx.progress_tracker.start()
-
-
-def _handle_preview_confirmation(
-    ctx: _MergeContext,
-    merge_results: list[MergeResult],
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]],
-    merged_count: int,
-    total_to_merge: int,
-) -> None:
-    """Handle the interactive preview-then-confirm flow.
-
-    Prompts the user for a continuation SHA and, if confirmed,
-    executes the real merge.
-    """
-    assert ctx.github_client is not None
-    assert ctx.source_pr is not None
-
-    console.print(f"\nMergeable {merged_count}/{total_to_merge} PRs")
-    # Per-PR preview lines are no longer printed during the run, so
-    # report the not-mergeable PRs (and why) here before prompting.
-    _print_failed_pr_details(merge_results)
-
-    if merged_count == 0:
-        console.print("\n\U0001f4a1 No PRs are mergeable at this time.")
-        return
-
-    commit_messages = ctx.github_client.get_pull_request_commits(
-        ctx.owner, ctx.repo_name, ctx.pr_number
-    )
-    first_commit_line = commit_messages[0].split("\n")[0] if commit_messages else ""
-    continue_sha_hash = _generate_continue_sha(ctx.source_pr, first_commit_line)
-    console.print()
-    console.print(f"To proceed with merging enter: {continue_sha_hash}")
-
-    try:
-        if "pytest" in sys.modules or os.getenv("TESTING"):
-            console.print("⚠️ Test mode detected - skipping interactive prompt")
-            return
-
-        user_input = input(
-            "Enter the string above to continue (or press Enter to cancel): "
-        ).strip()
-
-        if user_input == continue_sha_hash:
-            _execute_confirmed_merge(ctx, merge_results, all_prs_to_merge)
-        elif user_input == "":
-            console.print("❌ Merge cancelled by user.")
-        else:
-            console.print("❌ Invalid input. Merge cancelled.")
-    except KeyboardInterrupt:
-        console.print("\n❌ Merge cancelled by user.")
-    except EOFError:
-        console.print("\n❌ Merge cancelled.")
-
-
-def _execute_confirmed_merge(
-    ctx: _MergeContext,
-    preview_results: list[MergeResult],
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]],
-) -> None:
-    """Run the real merge after user confirmation."""
-    mergeable_prs = [
-        all_prs_to_merge[i]
-        for i, result in enumerate(preview_results)
-        if result.status.value == "merged"
-    ]
-    merged_count = len(mergeable_prs)
-
-    _restart_merge_progress_tracker(ctx, merged_count)
-    try:
-        real_results = _run_parallel_merge(ctx, mergeable_prs, preview=False)
-    finally:
-        if ctx.show_progress and ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-
-    _print_final_merge_summary(real_results)
-
-
-def _print_final_merge_summary(real_results: list[MergeResult]) -> None:
-    """Print the post-run 🚀 Final Results line and per-outcome recap.
-
-    Shared by the org / repo / similar-PR confirmed-merge paths so
-    every outcome category (including closed-without-merge) renders
-    identically regardless of scope.
-    """
-    final_merged = sum(1 for r in real_results if r.status.value == "merged")
-    final_failed = sum(1 for r in real_results if r.status.value == "failed")
-    final_skipped = sum(1 for r in real_results if r.status.value == "skipped")
-    final_blocked = sum(1 for r in real_results if r.status.value == "blocked")
-    final_closed = sum(1 for r in real_results if r.status.value == "closed")
-    final_auto_merge = sum(
-        1 for r in real_results if r.status.value == "auto_merge_pending"
-    )
-    parts = [f"{final_merged} merged"]
-    if final_auto_merge > 0:
-        parts.append(f"{final_auto_merge} auto-merge pending")
-    parts.append(f"{final_failed} failed")
-    if final_skipped > 0:
-        parts.append(f"{final_skipped} skipped")
-    if final_blocked > 0:
-        parts.append(f"{final_blocked} blocked")
-    if final_closed > 0:
-        parts.append(f"{final_closed} closed")
-    console.print(f"\n🚀 Final Results: {', '.join(parts)}")
-    if final_skipped > 0:
-        console.print(f"⏭️ Skipped {final_skipped} PRs")
-    if final_blocked > 0:
-        console.print(f"🛑 Blocked {final_blocked} PRs")
-    if final_closed > 0:
-        console.print(f"🚪 Closed without merging: {final_closed} PRs")
-    if final_auto_merge > 0:
-        console.print(f"⏳ Auto-merge pending for {final_auto_merge} PRs")
-
-    _print_failed_pr_details(real_results)
-
-
-def _format_failure_reason(reason: str) -> list[str]:
-    """Expand a failure reason into consistent display lines.
-
-    Failures are rendered in a consistent shape so the final summary is
-    actionable at a glance:
-
-      1. the failed PR URL (prepended by the caller),
-      2. a single **failure-type** line whose parts are joined with
-         `` / `` and carry no trailing colon, e.g.
-         ``Repository rule violations found / Required workflows failed``,
-      3. one bullet (``• ``) per individual failing condition.
-
-    Repository-ruleset violations arrive from GitHub as one long string
-    that crams the offending workflow / status-check names into a
-    quoted, comma-separated clause.  We split that into the type line
-    plus a bullet per name for both the ``Required workflows`` and
-    ``Required status check(s)`` variants.  Reasons we do not recognise
-    are returned unchanged as a single line.
-    """
-    if is_rule_violation(reason):
-        ruleset = RULE_VIOLATION_MARKER
-        verb = violation_verb(reason)
-        workflows = required_workflow_names(reason)
-        if workflows:
-            return [
-                f"{ruleset} / Required workflows {verb}",
-                *(f"• {name}" for name in workflows),
-            ]
-        checks = required_status_check_names(reason)
-        if checks:
-            return [
-                f"{ruleset} / Required status checks {verb}",
-                *(f"• {name}" for name in checks),
-            ]
-    return [reason]
-
-
-def _print_failed_pr_details(
-    merge_results: list[MergeResult],
-) -> None:
-    """Print URL and reason for every non-merged PR in the result list.
-
-    A bare ``Failed: 1`` line in the summary forces the user to
-    scroll back through the merge output to find which PR failed
-    and why.  During a real merge run the per-PR status lines are
-    no longer printed to the console at all (progress is conveyed
-    by the live tracker counters), so this end-of-run report is
-    the *only* place reasons appear.  It therefore covers every
-    non-merged terminal outcome — failed, blocked, skipped, closed
-    and auto-merge pending — one section per outcome.
-    """
-    sections: list[tuple[str, str]] = [
-        ("failed", "\n❌ Failed PRs:"),
-        ("blocked", "\n🛑 Blocked PRs:"),
-        ("skipped", "\n⏭️ Skipped PRs:"),
-        ("closed", "\n🚪 Closed PRs:"),
-        ("auto_merge_pending", "\n🤖 Auto-merge pending PRs:"),
-    ]
-    for status_value, heading in sections:
-        matching = [r for r in merge_results if r.status.value == status_value]
-        if not matching:
-            continue
-        console.print(heading)
-        for r in matching:
-            url = getattr(r.pr_info, "html_url", "<unknown>")
-            reason = r.error or "no reason reported"
-            body = "\n".join(f"     {line}" for line in _format_failure_reason(reason))
-            # markup=False so bracketed reasons are not eaten by Rich.
-            console.print(f"   • {url}\n{body}", markup=False)
-
-
-def _display_merge_results(
-    merge_results: list[MergeResult],
-    no_confirm: bool,
-) -> None:
-    """Print the final summary of merge results."""
-    merged_count = sum(1 for r in merge_results if r.status.value == "merged")
-    failed_count = sum(1 for r in merge_results if r.status.value == "failed")
-    skipped_count = sum(1 for r in merge_results if r.status.value == "skipped")
-    blocked_count = sum(1 for r in merge_results if r.status.value == "blocked")
-    closed_count = sum(1 for r in merge_results if r.status.value == "closed")
-    auto_merge_count = sum(
-        1 for r in merge_results if r.status.value == "auto_merge_pending"
-    )
-
-    if failed_count > 0:
-        if not no_confirm:
-            console.print(f"❌ Would fail to merge {failed_count} PRs")
-        else:
-            console.print(f"❌ Failed {failed_count} PRs")
-    if skipped_count > 0:
-        console.print(f"⏭️ Skipped {skipped_count} PRs")
-    if blocked_count > 0:
-        console.print(f"🛑 Blocked {blocked_count} PRs")
-    if closed_count > 0:
-        console.print(f"🚪 Closed without merging: {closed_count} PRs")
-    if auto_merge_count > 0:
-        console.print(f"⏳ Auto-merge pending for {auto_merge_count} PRs")
-
-    if no_confirm:
-        parts = [f"{merged_count} merged"]
-        if auto_merge_count > 0:
-            parts.append(f"{auto_merge_count} auto-merge pending")
-        parts.append(f"{failed_count} failed")
-        if skipped_count > 0:
-            parts.append(f"{skipped_count} skipped")
-        if blocked_count > 0:
-            parts.append(f"{blocked_count} blocked")
-        if closed_count > 0:
-            parts.append(f"{closed_count} closed")
-        console.print(f"📈 Final Results: {', '.join(parts)}")
-
-    _print_failed_pr_details(merge_results)
-
-
-def _handle_repo_merge(
-    parsed_repo: ParsedRepoUrl,
-    ctx: _MergeContext,
-) -> None:
-    """Handle merge operation for a repository-scoped URL.
-
-    Instead of scanning an entire org for similar PRs, this fetches all
-    open PRs in a single repository and merges the automation ones (or
-    all of them when --include-human-prs is given).
-
-    Args:
-        parsed_repo: Parsed repository URL with owner and repo.
-        ctx: Shared merge context populated with CLI parameters.
-    """
-    from .github_service import GitHubService
-    from .url_parser import _host_matches
-
-    if not _host_matches(parsed_repo.host, "github.com"):
-        console.print(
-            "❌ Repository-scoped merge is currently only supported "
-            f"for github.com (got host: {parsed_repo.host}).\n"
-            "   GitHub Enterprise support requires API base URL "
-            "configuration — use a direct PR URL instead."
-        )
-        raise typer.Exit(code=1)
-
-    ctx.github_client = GitHubClient(ctx.token)
-    assert ctx.github_client.token is not None
-    ctx.token = ctx.github_client.token
-    ctx.owner = parsed_repo.owner
-    ctx.repo_name = parsed_repo.repo
-
-    console.print(f"🔍 Repository mode: fetching open PRs in {parsed_repo.project}...")
-
-    _maybe_check_merge_permissions(ctx)
-
-    if ctx.show_progress:
-        ctx.progress_tracker = MergeProgressTracker(
-            f"{ctx.owner}/{ctx.repo_name}",
-            operation_label="Fetching open PRs",
-            operation_icon="🔍",
-        )
-        # Repo-scoped runs operate on a single repository, so the
-        # ``X/Y repos`` progress fraction is meaningless.  Skip
-        # ``update_total_repositories`` so the tracker falls through
-        # to the no-progress display branch and renders cleanly as
-        # ``🔍 Fetching open PRs in <owner>/<repo>``.
-        ctx.progress_tracker.start()
-
-    only_automation = not ctx.include_human_prs
-
-    async def _fetch_prs() -> list[PullRequestInfo]:
-        svc = GitHubService(
-            token=ctx.token,
-            progress_tracker=ctx.progress_tracker,
-        )
-        try:
-            return await svc.fetch_repo_open_prs(
-                ctx.owner,
-                ctx.repo_name,
-                only_automation=only_automation,
-            )
-        finally:
-            await svc.close()
-
-    try:
-        repo_prs = asyncio.run(_fetch_prs())
-    except Exception:
-        if ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-        raise
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-
-    if not repo_prs:
-        label = "automation " if only_automation else ""
-        console.print(f"❌ No open {label}PRs found in {parsed_repo.project}")
-        return
-
-    # The GraphQL fetch returns PRs newest-first (CREATED_AT DESC), but
-    # merging within a repository must drain oldest-first.  Merging the
-    # newest PR ahead of an older sibling leaves the older one behind the
-    # base branch, forcing automation (e.g. dependabot) into an avoidable
-    # rebase-and-revalidate cycle that can block the batch.  Sorting here
-    # mirrors the within-repository key applied owner-wide by
-    # ``_owner_merge_order`` so both schemes sequence a repository's PRs
-    # identically; the preview list below is derived from this order too.
-    repo_prs = _repo_merge_order(repo_prs)
-
-    automation_prs: list[PullRequestInfo] = []
-    human_prs: list[PullRequestInfo] = []
-
-    for pr in repo_prs:
-        # is_automation_author normalizes REST and GraphQL login forms
-        # (e.g. "dependabot[bot]" vs "dependabot") so they classify
-        # identically.
-        if is_automation_author(pr.author):
-            automation_prs.append(pr)
-        else:
-            human_prs.append(pr)
-
-    console.print(f"\n📊 Found {len(repo_prs)} open PR(s) in {parsed_repo.project}")
-    if automation_prs:
-        console.print(f"🤖 Automation PRs: {len(automation_prs)}")
-    if human_prs:
-        console.print(f"👤 Human PRs: {len(human_prs)}")
-
-    # List PRs that will be processed. The trailing ``(by {author})``
-    # already reveals whether each PR is automation or human, so a
-    # per-row icon would only duplicate the summary counts above.
-    for pr in repo_prs:
-        console.print(f"  #{pr.number} {pr.title} (by {pr.author})")
-
-    # Only prompt when human PRs are actually in scope, not merely
-    # because --include-human-prs was supplied.  A dry run never prompts;
-    # it mirrors the safe default (exclude human PRs) so the preview
-    # matches a real run where the operator presses Enter at the prompt.
-    needs_human_confirm = bool(human_prs) and not ctx.no_confirm and not ctx.dry_run
-    if needs_human_confirm:
-        console.print("\n⚠️ Human-authored PRs are included in this merge operation.")
-        console.print("   Review the list above carefully before proceeding.")
-        try:
-            user_input = (
-                typer.prompt(
-                    "Type 'yes' to include human PRs, or press Enter to skip them",
-                    default="",
-                    show_default=False,
-                )
-                .strip()
-                .lower()
-            )
-            if user_input != "yes":
-                console.print("ℹ️ Excluding human PRs from merge.")
-                repo_prs = automation_prs
-                if not repo_prs:
-                    console.print("❌ No automation PRs remain to merge.")
-                    return
-        except (KeyboardInterrupt, EOFError, typer.Abort):
-            console.print("\n❌ Merge cancelled by user.")
-            return
-    elif human_prs and ctx.dry_run:
-        # Human PRs only reach this branch when --include-human-prs was
-        # supplied (otherwise they are filtered out at fetch time).  A dry
-        # run performs no writes, so keep them in the preview to faithfully
-        # mirror what a real --include-human-prs run would attempt; a real
-        # run would prompt for confirmation first (unless --no-confirm).
-        console.print(
-            "\nℹ️ Dry run: human-authored PRs are kept in this preview "
-            "(a real run would prompt before merging them)."
-        )
-
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]] = [
-        (pr, None) for pr in repo_prs
-    ]
-
-    # Without --no-confirm this first pass is a preview (evaluation)
-    # run — label the tracker accordingly so its counters ("Mergeable")
-    # don't claim merges that never happened.
-    preview_run = ctx.dry_run or not ctx.no_confirm
-    if ctx.show_progress:
-        ctx.progress_tracker = MergeProgressTracker(
-            f"{ctx.owner}/{ctx.repo_name}",
-            operation_label="Evaluating PRs" if preview_run else "Merging PRs",
-            operation_icon="\U0001f50d" if preview_run else "\u25b6\ufe0f",
-            preview=preview_run,
-        )
-        ctx.progress_tracker.set_total_prs(len(all_prs_to_merge))
-        ctx.progress_tracker.start()
-
-    try:
-        # Per-repo merge dispatch is serialised inside
-        # ``AsyncMergeManager`` (see ``_get_merge_dispatch_lock``),
-        # so it is now safe to run multiple workers against PRs
-        # that target the same repository — only the actual
-        # ``merge_pull_request`` call queues, while approve,
-        # rebase, and the Step 5.5 auto-merge wait run in
-        # parallel.
-        merge_results = _run_parallel_merge(
-            ctx,
-            all_prs_to_merge,
-            preview=preview_run,
-            # Allow parallel workers; the merge dispatch itself is
-            # serialised per repo by ``AsyncMergeManager`` so PRs
-            # parked in Step 5.5's wait loop no longer block other
-            # PRs in the batch.  Cap by PR count so we don't spawn
-            # more workers than there is work.
-            concurrency=min(5, len(all_prs_to_merge)) or 1,
-            # All PRs target the same repo, so a sibling merge can make
-            # a queued PR ``dirty`` / ``behind`` mid-batch; refresh
-            # live state before each merge dispatch.
-            repo_scoped=True,
-        )
-    finally:
-        if ctx.show_progress and ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-
-    if not merge_results:
-        console.print("❌ No PRs were processed")
-        return
-
-    merged_count = sum(1 for r in merge_results if r.status.value == "merged")
-
-    # Dry run: report the preview and stop before any prompt or merge.
-    if ctx.dry_run:
-        _display_merge_results(merge_results, no_confirm=False)
-        return
-
-    if not ctx.no_confirm:
-        # In preview mode, show what would happen, then prompt
-        # for confirmation via an override-style SHA.
-        _handle_repo_preview_confirmation(
-            ctx,
-            merge_results,
-            all_prs_to_merge,
-            merged_count,
-            len(merge_results),
-        )
-        return
-
-    _display_merge_results(merge_results, ctx.no_confirm)
-
-
-def _handle_repo_preview_confirmation(
-    ctx: _MergeContext,
-    merge_results: list[MergeResult],
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]],
-    merged_count: int,
-    total_to_merge: int,
-) -> None:
-    """Handle preview-then-confirm for repository-scoped merges.
-
-    Similar to _handle_preview_confirmation but does not require a
-    source PR for SHA generation — it uses the repository name instead.
-    """
-    console.print(f"\nMergeable {merged_count}/{total_to_merge} PRs")
-    # Per-PR preview lines are no longer printed during the run, so
-    # report the not-mergeable PRs (and why) here before prompting.
-    _print_failed_pr_details(merge_results)
-
-    if merged_count == 0:
-        console.print("\n\U0001f4a1 No PRs are mergeable at this time.")
-        return
-
-    # Generate a confirmation token from the repo context
-    combined = f"repo-merge:{ctx.owner}/{ctx.repo_name}:{merged_count}"
-    confirm_hash = hashlib.sha256(combined.encode("utf-8")).hexdigest()[:16]
-
-    console.print()
-    console.print(f"To proceed with merging enter: {confirm_hash}")
-
-    try:
-        user_input = typer.prompt(
-            "Enter the string above to continue (or press Enter to cancel)",
-            default="",
-            show_default=False,
-        ).strip()
-
-        if user_input == confirm_hash:
-            _execute_repo_confirmed_merge(ctx, merge_results, all_prs_to_merge)
-        elif user_input == "":
-            console.print("❌ Merge cancelled by user.")
-        else:
-            console.print("❌ Invalid input. Merge cancelled.")
-    except (KeyboardInterrupt, EOFError, typer.Abort):
-        console.print("\n❌ Merge cancelled by user.")
-
-
-def _execute_repo_confirmed_merge(
-    ctx: _MergeContext,
-    preview_results: list[MergeResult],
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]],
-) -> None:
-    """Run the real merge after user confirmation (repo mode)."""
-    mergeable_prs = [
-        all_prs_to_merge[i]
-        for i, result in enumerate(preview_results)
-        if result.status.value == "merged"
-    ]
-
-    if ctx.show_progress:
-        ctx.progress_tracker = MergeProgressTracker(
-            f"{ctx.owner}/{ctx.repo_name}",
-            operation_label="Merging PRs",
-            operation_icon="▶️",
-        )
-        ctx.progress_tracker.set_total_prs(len(mergeable_prs))
-        ctx.progress_tracker.start()
-
-    try:
-        real_results = _run_parallel_merge(
-            ctx,
-            mergeable_prs,
-            preview=False,
-            # Per-repo merge dispatch lock makes parallel workers
-            # safe; cap by PR count.
-            concurrency=min(5, len(mergeable_prs)) or 1,
-            # Single-repo batch — refresh live merge state before each
-            # dispatch (a sibling merge can introduce a conflict).
-            repo_scoped=True,
-        )
-    finally:
-        if ctx.show_progress and ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-
-    _print_final_merge_summary(real_results)
-
-
-def _repo_merge_order(
-    prs: list[PullRequestInfo],
-) -> list[PullRequestInfo]:
-    """Order a single repository's PRs for oldest-first merging.
-
-    Sorts ascending by PR number, i.e. in the order the automation raised
-    them (oldest first).  Merging the oldest PR first minimises the rebase
-    churn imposed on the newer siblings: each merge advances the base
-    branch, so a newer sibling merged ahead of an older one would leave
-    the older PR ``behind`` and trigger an avoidable rebase + CI wait.
-
-    This is the single-repository analogue of the within-repository key
-    used owner-wide by :func:`_owner_merge_order`, keeping both schemes'
-    intra-repository sequencing identical.
-    """
-    return sorted(prs, key=lambda p: p.number)
-
-
-def _owner_merge_order(
-    prs: list[PullRequestInfo],
-) -> list[PullRequestInfo]:
-    """Order owner-wide PRs for striped merging.
-
-    Sorts so that:
-
-    - Repositories with the most in-scope PRs come first.  These take the
-      longest to drain (each merge can make the next sibling ``behind`` /
-      ``dirty`` and trigger a rebase + CI wait), so starting them earliest
-      gives them the most wall-clock head start under the striped
-      scheduler's concurrent per-repository workers.
-    - Within a repository, PRs ascend by number, i.e. in the order the
-      automation raised them (oldest first).  Merging the oldest first
-      minimises the rebase churn imposed on the newer siblings.
-
-    Ties between equally-sized repositories break on repository name so
-    the order (and the grouped listing derived from it) is deterministic.
-    """
-    counts: dict[str, int] = {}
-    for pr in prs:
-        counts[pr.repository_full_name] = counts.get(pr.repository_full_name, 0) + 1
-    return sorted(
-        prs,
-        key=lambda p: (
-            -counts[p.repository_full_name],
-            p.repository_full_name,
-            p.number,
-        ),
-    )
-
-
-def _print_prs_grouped_by_repo(
-    prs: list[PullRequestInfo],
-) -> None:
-    """Print a PR list grouped by repository for owner-wide readability.
-
-    Emits a header per repository followed by its PRs indented beneath,
-    so a large owner-wide list stays scannable.  Repositories are listed
-    in the order they first appear in ``prs`` (the caller passes them in
-    merge order via :func:`_owner_merge_order`, so the listing mirrors the
-    sequence in which they will be merged); PRs within a repository are
-    shown in the supplied order.
-    """
-    by_repo: dict[str, list[PullRequestInfo]] = {}
-    for pr in prs:
-        by_repo.setdefault(pr.repository_full_name, []).append(pr)
-
-    for repo, repo_prs in by_repo.items():
-        console.print(f"\n📁 {repo} ({len(repo_prs)} PR(s))")
-        for pr in repo_prs:
-            console.print(f"  #{pr.number} {pr.title} (by {pr.author})")
-
-
-def _handle_org_merge(
-    parsed_org: ParsedOrgUrl,
-    ctx: _MergeContext,
-) -> None:
-    """Handle merge operation for an owner-scoped (org/user) URL.
-
-    Enumerates every in-scope automation PR across all of the owner's
-    non-archived, non-fork repositories, then bulk merges them with the
-    striped scheduler so no two merges stack on the same repository at
-    once.  The owner may be an organization or a personal user account;
-    the account type is detected at runtime during enumeration.
-
-    Args:
-        parsed_org: Parsed owner URL with the org/user login.
-        ctx: Shared merge context populated with CLI parameters.
-    """
-    from .github_service import GitHubService
-
-    # The host here is guaranteed to *match* github.com: ``parse_org_url``
-    # is the single github.com-only choke point (it rejects every other
-    # host, including GHE, before a ``ParsedOrgUrl`` can reach this
-    # handler).  "Matches" is deliberate — ``parse_org_url`` accepts
-    # github.com and its subdomains (e.g. ``api.github.com``) via
-    # ``_host_matches``, so this is not an exact ``== "github.com"``
-    # guarantee.  Enabling GHE (#343) means relaxing that one parser
-    # guard and threading ``derive_api_urls(host)`` through the service
-    # stack — deliberately not re-checked here so there is no second
-    # guard to drift out of sync.
-
-    ctx.github_client = GitHubClient(ctx.token)
-    assert ctx.github_client.token is not None
-    ctx.token = ctx.github_client.token
-    ctx.owner = parsed_org.owner
-    ctx.repo_name = ""
-
-    console.print(f"🔍 Owner mode: scanning {parsed_org.owner} for automation PRs...")
-
-    # NOTE: the token permission check is deferred until *after*
-    # enumeration (see below).  ``_check_merge_permissions`` probes a
-    # concrete repository, and ``check_token_permissions`` reports every
-    # operation as missing when no repo is supplied — running it here
-    # with an empty ``ctx.repo_name`` would abort every owner-wide run
-    # unconditionally.
-
-    if ctx.show_progress:
-        ctx.progress_tracker = MergeProgressTracker(
-            parsed_org.owner,
-            operation_label="Scanning repositories for automation PRs",
-            operation_icon="🔍",
-        )
-        ctx.progress_tracker.start()
-
-    only_automation = not ctx.include_human_prs
-
-    async def _fetch_prs() -> tuple[list[PullRequestInfo], list[str]]:
-        svc = GitHubService(
-            token=ctx.token,
-            progress_tracker=ctx.progress_tracker,
-        )
-        try:
-            return await svc.fetch_owner_open_prs(
-                parsed_org.owner,
-                only_automation=only_automation,
-            )
-        finally:
-            await svc.close()
-
-    try:
-        owner_prs, scan_errors = asyncio.run(_fetch_prs())
-    except Exception:
-        if ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-        raise
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-
-    if scan_errors:
-        error_count = len(scan_errors)
-        repo_noun = "repository" if error_count == 1 else "repositories"
-        console.print(f"\n⚠️ {error_count} {repo_noun} could not be scanned:")
-        for err in scan_errors:
-            console.print(f"   - {err}")
-
-    if not owner_prs:
-        label = "automation " if only_automation else ""
-        console.print(f"❌ No open {label}PRs found in {parsed_org.owner}")
-        return
-
-    # Order for striped merging: repositories with the most PRs first,
-    # ascending PR number within each repository (see _owner_merge_order).
-    # Both the grouped listing and the merge list derive from this order.
-    owner_prs = _owner_merge_order(owner_prs)
-
-    # Deferred from before enumeration: the helper needs a concrete repo
-    # to probe, so point it at the first in-scope PR's repository as a
-    # representative sample.  The common failure modes (expired/invalid
-    # token, no write access) fail identically on any repo; genuine
-    # per-repo permission variance with fine-grained tokens is caught at
-    # merge time by the scheduler's per-repository error isolation.
-    ctx.repo_name = owner_prs[0].repository_full_name.split("/", 1)[-1]
-    _maybe_check_merge_permissions(ctx)
-
-    automation_prs: list[PullRequestInfo] = []
-    human_prs: list[PullRequestInfo] = []
-    for pr in owner_prs:
-        if is_automation_author(pr.author):
-            automation_prs.append(pr)
-        else:
-            human_prs.append(pr)
-
-    distinct_repos = {pr.repository_full_name for pr in owner_prs}
-    repo_count = len(distinct_repos)
-    repo_noun = "repository" if repo_count == 1 else "repositories"
-    console.print(
-        f"\n📊 Found {len(owner_prs)} open PR(s) across "
-        f"{repo_count} {repo_noun} in {parsed_org.owner}"
-    )
-    if automation_prs:
-        console.print(f"🤖 Automation PRs: {len(automation_prs)}")
-    if human_prs:
-        console.print(f"👤 Human PRs: {len(human_prs)}")
-
-    # Grouped-by-repository listing keeps a large owner-wide list scannable.
-    _print_prs_grouped_by_repo(owner_prs)
-
-    needs_human_confirm = bool(human_prs) and not ctx.no_confirm and not ctx.dry_run
-    if needs_human_confirm:
-        console.print(
-            "\n⚠️ Human-authored PRs across the entire owner "
-            f"'{parsed_org.owner}' are included in this merge operation."
-        )
-        console.print("   Review the list above carefully before proceeding.")
-        try:
-            user_input = (
-                typer.prompt(
-                    "Type 'yes' to include human PRs, or press Enter to skip them",
-                    default="",
-                    show_default=False,
-                )
-                .strip()
-                .lower()
-            )
-            if user_input != "yes":
-                console.print("ℹ️ Excluding human PRs from merge.")
-                owner_prs = automation_prs
-                if not owner_prs:
-                    console.print("❌ No automation PRs remain to merge.")
-                    return
-        except (KeyboardInterrupt, EOFError, typer.Abort):
-            console.print("\n❌ Merge cancelled by user.")
-            return
-    elif human_prs and ctx.dry_run:
-        # Human PRs only reach this branch when --include-human-prs was
-        # supplied (otherwise they are filtered out at fetch time).  A dry
-        # run performs no writes, so keep them in the preview to faithfully
-        # mirror what a real --include-human-prs run would attempt; a real
-        # run would prompt for confirmation first (unless --no-confirm).
-        console.print(
-            "\nℹ️ Dry run: human-authored PRs are kept in this preview "
-            "(a real run would prompt before merging them)."
-        )
-
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]] = [
-        (pr, None) for pr in owner_prs
-    ]
-
-    # Global concurrency 10, bounded by the distinct-repo count: the
-    # striped scheduler runs at most one PR per repo, so more workers
-    # than repositories cannot help.
-    final_distinct_repos = {pr.repository_full_name for pr in owner_prs}
-    concurrency = min(10, len(final_distinct_repos)) or 1
-
-    # Without --no-confirm this first pass is a preview (evaluation)
-    # run — label the tracker accordingly so its counters ("Mergeable")
-    # don't claim merges that never happened.
-    preview_run = ctx.dry_run or not ctx.no_confirm
-    if ctx.show_progress:
-        ctx.progress_tracker = MergeProgressTracker(
-            parsed_org.owner,
-            operation_label="Evaluating PRs" if preview_run else "Merging PRs",
-            operation_icon="\U0001f50d" if preview_run else "\u25b6\ufe0f",
-            preview=preview_run,
-        )
-        ctx.progress_tracker.set_total_prs(len(all_prs_to_merge))
-        ctx.progress_tracker.start()
-
-    try:
-        merge_results = _run_parallel_merge(
-            ctx,
-            all_prs_to_merge,
-            preview=preview_run,
-            concurrency=concurrency,
-            # Owner-wide batches mix many repositories and often contain
-            # multiple PRs per repository; stripe to avoid stacking and
-            # repo_scoped to refresh live merge state before each dispatch.
-            repo_scoped=True,
-            stripe=True,
-        )
-    finally:
-        if ctx.show_progress and ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-
-    if not merge_results:
-        console.print("❌ No PRs were processed")
-        return
-
-    merged_count = sum(1 for r in merge_results if r.status.value == "merged")
-
-    # Dry run: report the preview and stop before any prompt or merge.
-    if ctx.dry_run:
-        _display_merge_results(merge_results, no_confirm=False)
-        return
-
-    if not ctx.no_confirm:
-        _handle_org_preview_confirmation(
-            ctx,
-            parsed_org,
-            merge_results,
-            all_prs_to_merge,
-            merged_count,
-            len(merge_results),
-            concurrency,
-        )
-        return
-
-    _display_merge_results(merge_results, ctx.no_confirm)
-
-
-def _handle_org_preview_confirmation(
-    ctx: _MergeContext,
-    parsed_org: ParsedOrgUrl,
-    merge_results: list[MergeResult],
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]],
-    merged_count: int,
-    total_to_merge: int,
-    concurrency: int,
-) -> None:
-    """Handle preview-then-confirm for owner-scoped merges.
-
-    Mirrors :func:`_handle_repo_preview_confirmation` but derives the
-    confirmation token from the owner login instead of a repository.
-    """
-    console.print(f"\nMergeable {merged_count}/{total_to_merge} PRs")
-    # Per-PR preview lines are no longer printed during the run, so
-    # report the not-mergeable PRs (and why) here before prompting.
-    _print_failed_pr_details(merge_results)
-
-    if merged_count == 0:
-        console.print("\n\U0001f4a1 No PRs are mergeable at this time.")
-        return
-
-    combined = f"org-merge:{parsed_org.owner}:{merged_count}"
-    confirm_hash = hashlib.sha256(combined.encode("utf-8")).hexdigest()[:16]
-
-    console.print()
-    console.print(f"To proceed with merging enter: {confirm_hash}")
-
-    try:
-        user_input = typer.prompt(
-            "Enter the string above to continue (or press Enter to cancel)",
-            default="",
-            show_default=False,
-        ).strip()
-
-        if user_input == confirm_hash:
-            _execute_org_confirmed_merge(
-                ctx, merge_results, all_prs_to_merge, concurrency
-            )
-        elif user_input == "":
-            console.print("❌ Merge cancelled by user.")
-        else:
-            console.print("❌ Invalid input. Merge cancelled.")
-    except (KeyboardInterrupt, EOFError, typer.Abort):
-        console.print("\n❌ Merge cancelled by user.")
-
-
-def _execute_org_confirmed_merge(
-    ctx: _MergeContext,
-    preview_results: list[MergeResult],
-    all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]],
-    concurrency: int,
-) -> None:
-    """Run the real owner-wide merge after user confirmation."""
-    mergeable_prs = [
-        all_prs_to_merge[i]
-        for i, result in enumerate(preview_results)
-        if result.status.value == "merged"
-    ]
-
-    if ctx.show_progress:
-        ctx.progress_tracker = MergeProgressTracker(
-            ctx.owner,
-            operation_label="Merging PRs",
-            operation_icon="▶️",
-        )
-        ctx.progress_tracker.set_total_prs(len(mergeable_prs))
-        ctx.progress_tracker.start()
-
-    try:
-        real_results = _run_parallel_merge(
-            ctx,
-            mergeable_prs,
-            preview=False,
-            concurrency=concurrency,
-            repo_scoped=True,
-            stripe=True,
-        )
-    finally:
-        if ctx.show_progress and ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-
-    _print_final_merge_summary(real_results)
-
-
-def _print_gerrit_final_summary(
-    results: list[GerritSubmitResult],
-    all_changes: list[tuple[GerritChangeInfo, GerritComparisonResult | None]],
-    console: Console,
-) -> None:
-    """Print the post-run 🚀 Final Results line and failure recap.
-
-    Mirrors the GitHub ``_print_final_merge_summary`` /
-    ``_print_failed_pr_details`` pair: per-change status lines are
-    not printed while the submission runs (progress is conveyed by
-    the live tracker counters), so this end-of-run report is the
-    only place failure reasons appear.
-
-    Args:
-        results: Submit results, one per attempted change.
-        all_changes: The (change, comparison) tuples that were
-            submitted; used to recover each change's web URL for the
-            failure recap (``GerritSubmitResult`` carries only the
-            project and number).
-        console: Rich console to print to.
-    """
-    submitted = sum(1 for r in results if r.submitted)
-    failed = [r for r in results if not r.success]
-    reviewed_only = sum(1 for r in results if r.reviewed and not r.submitted)
-
-    parts = [f"{submitted} submitted", f"{len(failed)} failed"]
-    if reviewed_only > 0:
-        parts.append(f"{reviewed_only} reviewed (not submitted)")
-    console.print(f"\n🚀 Final Results: {', '.join(parts)}")
-
-    if not failed:
-        return
-    url_by_key = {
-        (change.project, change.number): change.url for change, _ in all_changes
-    }
-    console.print("\n❌ Failed changes:")
-    for result in failed:
-        url = url_by_key.get((result.project, result.change_number)) or (
-            f"{result.project} #{result.change_number}"
-        )
-        reason = result.error or "no reason reported"
-        # markup=False so bracketed reasons are not eaten by Rich.
-        console.print(f"   • {url}\n     {reason}", markup=False)
-
-
-def _resolve_gerrit_credentials_or_exit(
-    parsed_url: ParsedUrl | ParsedGerritTopicUrl,
-    no_netrc: bool,
-    netrc_file: Path | None,
-    verbose: bool,
-    console: Console,
-) -> GerritCredentials:
-    """Resolve Gerrit credentials, or print guidance and exit.
-
-    Args:
-        parsed_url: Parsed Gerrit change or topic URL (for the host).
-        no_netrc: If True, skip .netrc credential lookup.
-        netrc_file: Explicit path to a .netrc file.
-        verbose: When True, report which auth method was used.
-        console: Rich console for output.
-
-    Returns:
-        Valid Gerrit credentials.
-
-    Raises:
-        typer.Exit: When no valid credentials can be resolved.
-    """
-    try:
-        credentials = resolve_gerrit_credentials(
-            host=parsed_url.host,
-            use_netrc=not no_netrc,
-            netrc_file=netrc_file,
-        )
-    except NetrcParseError as e:
-        console.print(f"⚠️ Error parsing .netrc file: {e}")
-        credentials = None
-
-    if credentials is None or not credentials.is_valid:
-        console.print("❌ Gerrit credentials not found.")
-        console.print("   Options:")
-        console.print("   1. Create a ~/.netrc file with Gerrit credentials")
-        console.print(
-            "   2. Set GERRIT_USERNAME and GERRIT_PASSWORD environment variables"
-        )
-        console.print(
-            "   Tip: Source your .secrets.gerrit file and run use_lf or use_onap"
-        )
-        raise typer.Exit(1)
-
-    if verbose:
-        console.print(f"🔑 Using credentials from {credentials.auth_method_display()}")
-
-    return credentials
-
-
-def _resolve_gerrit_source_change(
-    service: GerritService,
-    parsed_url: ParsedUrl | ParsedGerritTopicUrl,
-    topic: str | None,
-    credentials: GerritCredentials,
-    console: Console,
-) -> tuple[GerritChangeInfo, list[GerritChangeInfo] | None]:
-    """Fetch and display the source change, validating its state.
-
-    For a topic search URL there is no explicit change number: the
-    first open change in the topic anchors the batch and the rest
-    become candidates.
-
-    Returns:
-        The source change and, when a topic URL was given, the list of
-        open changes sharing that topic (else None).
-
-    Raises:
-        typer.Exit: When no change is found, or the source change is
-            already merged or abandoned.
-    """
-    topic_changes: list[GerritChangeInfo] | None = None
-    if isinstance(parsed_url, ParsedGerritTopicUrl):
-        search_topic = topic or parsed_url.topic
-        console.print(f"📋 Fetching changes with topic '{search_topic}'...")
-        topic_changes = [
-            c for c in service.get_changes_by_topic(search_topic) if c.is_open
-        ]
-        if not topic_changes:
-            console.print(f"❌ No open changes found with topic '{search_topic}'")
-            raise typer.Exit(1)
-        # Refetch the anchor change: list queries omit the label,
-        # permission, and action detail the checks below rely on.
-        source_change = service.get_change_info(topic_changes[0].number)
-    else:
-        console.print(f"📋 Fetching change {parsed_url.change_number}...")
-        source_change = service.get_change_info(parsed_url.change_number)
-
-    if source_change is None:
-        console.print("❌ Change not found")
-        raise typer.Exit(1)
-
-    # Display source change info using Rich table (same style as GitHub)
-    _display_change_info(
-        source_change,
-        console=console,
-        auth_method=credentials.auth_method_display(),
-    )
-
-    if source_change.status == "MERGED":
-        console.print("\n✅ Change is already merged.")
-        raise typer.Exit(0)
-
-    if source_change.status == "ABANDONED":
-        console.print("\n❌ Change has been abandoned.")
-        raise typer.Exit(1)
-
-    return source_change, topic_changes
-
-
-def _resolve_gerrit_only_automation(
-    source_change: GerritChangeInfo,
-    comparator: GerritChangeComparator,
-    override: str | None,
-    console: Console,
-) -> bool:
-    """Decide whether the batch is automation-only, honouring --override.
-
-    Automation source changes match only other automation changes. A
-    non-automation source requires a matching override SHA to proceed
-    and then widens matching beyond automation.
-
-    Returns:
-        True when only automation changes should be matched.
-
-    Raises:
-        SystemExit: When the source is a non-automation change without a
-            valid override SHA.
-    """
-    if comparator.is_automation_change(source_change):
-        return True
-
-    expected_sha = _generate_gerrit_override_sha(source_change)
-    if not override:
-        owner = source_change.owner.strip()
-        subject = source_change.subject.strip()
-        subject_preview = subject if len(subject) <= 50 else f"{subject[:50]}..."
-        console.print("Source change is not from a recognized automation tool.")
-        console.print(
-            "To submit this and similar changes, run again with: "
-            f"--override {expected_sha}"
-        )
-        console.print(
-            f"This SHA is based on the owner '{owner}' and subject '{subject_preview}'",
-            style="dim",
-        )
-        raise typer.Exit(0)
-
-    if override.strip().lower() != expected_sha:
-        exit_with_error(
-            ExitCode.VALIDATION_ERROR,
-            message="❌ Invalid override SHA provided",
-            details=(
-                f"Expected SHA for this change and owner: --override {expected_sha}"
-            ),
-        )
-
-    console.print(
-        "Override SHA validated. Proceeding with non-automation change merge."
-    )
-    return False
-
-
-def _maybe_rebase_gerrit_change(
-    service: GerritService,
-    source_change: GerritChangeInfo,
-    credentials: GerritCredentials,
-    console: Console,
-) -> GerritChangeInfo:
-    """Rebase the source change when it has merge conflicts.
-
-    Returns:
-        The (possibly refreshed) source change. Unchanged when the
-        change is already mergeable.
-
-    Raises:
-        typer.Exit: When the rebase fails, whether due to conflicts
-            needing manual resolution or another error.
-    """
-    if source_change.mergeable is not False:
-        return source_change
-
-    console.print("\n⚠️ Change has merge conflicts. Attempting to rebase...")
-    rebase_result = service.rebase_change(source_change.number)
-
-    if rebase_result["success"]:
-        console.print("✅ Rebase successful! Refreshing change info...")
-        source_change = service.get_change_info(source_change.number)
-        _display_change_info(
-            source_change,
-            console=console,
-            auth_method=credentials.auth_method_display(),
-        )
-        return source_change
-
-    if rebase_result["conflict"]:
-        console.print("\n❌ Rebase failed due to merge conflicts:")
-        if rebase_result["conflicting_files"]:
-            console.print("\n   Conflicting files:")
-            for file_path in rebase_result["conflicting_files"]:
-                console.print(f"   • {file_path}")
-        console.print(
-            "\n💡 To resolve: manually rebase the change locally and push a new patchset."
-        )
-        console.print(f"   git review -d {source_change.number}")
-        console.print(f"   git rebase origin/{source_change.branch}")
-        console.print("   # resolve conflicts, then:")
-        console.print("   git review")
-        raise typer.Exit(1)
-
-    console.print(f"\n❌ Rebase failed: {rebase_result['error']}")
-    raise typer.Exit(1)
-
-
-def _resolve_gerrit_candidates(
-    service: GerritService,
-    source_change: GerritChangeInfo,
-    parsed_url: ParsedUrl | ParsedGerritTopicUrl,
-    topic: str | None,
-    topic_changes: list[GerritChangeInfo] | None,
-    console: Console,
-) -> list[GerritChangeInfo] | None:
-    """Resolve the candidate changes to compare against, by topic.
-
-    An explicit --topic wins, then the topic from a search URL, then
-    the source change's own topic. A server-side topic query is far
-    cheaper and more reliable than scanning every open change; when no
-    topic is available, None lets the caller fall back to a full scan.
-    """
-    effective_topic = topic or source_change.topic
-    if isinstance(parsed_url, ParsedGerritTopicUrl):
-        effective_topic = topic or parsed_url.topic
-
-    if not effective_topic:
-        console.print(f"\n🔍 Searching for similar changes on {parsed_url.host}...")
-        return None
-
-    console.print(
-        f"\n🔍 Searching for changes with topic "
-        f"'{effective_topic}' on {parsed_url.host}..."
-    )
-    # topic_changes (when set) was fetched with this same topic
-    if topic_changes is not None:
-        return topic_changes
-    return [c for c in service.get_changes_by_topic(effective_topic) if c.is_open]
-
-
-def _find_and_print_similar_changes(
-    service: GerritService,
-    comparator: GerritChangeComparator,
-    source_change: GerritChangeInfo,
-    candidates: list[GerritChangeInfo] | None,
-    only_automation: bool,
-    console: Console,
-) -> list[tuple[GerritChangeInfo, GerritComparisonResult]]:
-    """Score candidates against the source change and print the matches."""
-    similar_changes = service.find_similar_changes(
-        source_change,
-        comparator,
-        only_automation=only_automation,
-        candidates=candidates,
-    )
-
-    console.print(f"Found {len(similar_changes)} similar changes:")
-    for change, comparison in similar_changes:
-        console.print(f"  • {change.project} #{change.number}: {change.subject}")
-        console.print(f"    {_format_gerrit_similarity(comparison)}")
-
-    return similar_changes
-
-
-def _preview_gerrit_submission(
-    source_change: GerritChangeInfo,
-    all_changes: list[tuple[GerritChangeInfo, GerritComparisonResult | None]],
-    no_confirm: bool,
-    dry_run: bool,
-    console: Console,
-) -> bool:
-    """Warn about permissions and preview the run, then decide to proceed.
-
-    Permissions are per-project in Gerrit, so this checks the source
-    change and warns if the user may lack sufficient permissions. It
-    then either stops (dry run), prompts for confirmation
-    (interactive), or proceeds straight to submission (--no-confirm).
-
-    Returns:
-        True when submission should proceed, False when the caller
-        should stop (dry run, or the user declined confirmation).
-    """
-    permission_warnings = source_change.get_permission_warnings()
-    if permission_warnings:
-        console.print("\n⚠️ Permission warnings:")
-        for warning in permission_warnings:
-            console.print(f"   • {warning}")
-        console.print(
-            "\n   Note: Permissions vary by project. The operation may still "
-            "succeed on some changes."
-        )
-
-    if not no_confirm or dry_run:
-        label = "Dry run" if dry_run else "Preview"
-        console.print(
-            f"\n📊 {label}: {len(all_changes)} changes would be reviewed and submitted"
-        )
-        if source_change.has_required_permissions():
-            console.print(
-                "   ✅ You appear to have required permissions (+2 Code-Review, submit)"
-            )
-        else:
-            console.print(
-                "   ⚠️ You may not have all required permissions (see warnings above)"
-            )
-        if dry_run:
-            console.print("\n🧪 Dry run: no changes were reviewed or submitted.")
-            return False
-        if not _confirm_gerrit_submission(source_change, console):
-            return False
-
-    return True
-
-
-def _run_gerrit_submission(
-    parsed_url: ParsedUrl | ParsedGerritTopicUrl,
-    credentials: GerritCredentials,
-    all_changes: list[tuple[GerritChangeInfo, GerritComparisonResult | None]],
-    show_progress: bool,
-    console: Console,
-) -> None:
-    """Submit all changes in parallel and print the final summary.
-
-    Live in-place progress mirrors the GitHub merge path: the submit
-    manager records each change's transitory and terminal states
-    against the tracker while the parallel submission runs, and
-    failures are recapped afterwards by _print_gerrit_final_summary
-    instead of interleaved lines. The tracker is created (unstarted)
-    before the submit manager so it can be handed over, but only
-    started inside the try/finally so it is always stopped, even when
-    submission setup or the run itself raises.
-    """
-    console.print(f"\n🚀 Submitting {len(all_changes)} changes...")
-
-    progress_tracker: MergeProgressTracker | None = None
-    if show_progress:
-        progress_tracker = MergeProgressTracker(
-            parsed_url.host,
-            operation_label="Submitting changes",
-            operation_icon="▶️",
-            unit_label="changes",
-        )
-        progress_tracker.set_total_prs(len(all_changes))
-
-    submit_manager = create_submit_manager(
-        host=parsed_url.host,
-        base_path=parsed_url.base_path,
-        username=credentials.username,
-        password=credentials.password,
-        progress_tracker=progress_tracker,
-    )
-
-    try:
-        if progress_tracker is not None:
-            progress_tracker.start()
-        results = submit_manager.submit_changes_parallel(all_changes)
-    finally:
-        if progress_tracker is not None:
-            progress_tracker.stop()
-
-    _print_gerrit_final_summary(results, all_changes, console)
-
-
-def _handle_gerrit_merge(
-    parsed_url: ParsedUrl | ParsedGerritTopicUrl,
-    no_confirm: bool,
-    similarity_threshold: float,
-    verbose: bool,
-    console: Console,
-    no_netrc: bool = False,
-    netrc_file: Path | None = None,
-    netrc_optional: bool = True,
-    dry_run: bool = False,
-    override: str | None = None,
-    topic: str | None = None,
-    show_progress: bool = True,
-) -> None:
-    """
-    Handle merge operation for a Gerrit change or topic search URL.
-
-    Args:
-        parsed_url: Parsed Gerrit change URL (host, project, change
-            number) or topic search URL (host, topic).
-        no_confirm: If True, skip confirmation prompt.
-        similarity_threshold: Threshold for matching similar changes.
-        verbose: Enable verbose output.
-        console: Rich console for output.
-        no_netrc: If True, skip .netrc credential lookup.
-        netrc_file: Explicit path to a .netrc file.
-        netrc_optional: If True, don't fail if netrc not found.
-        dry_run: If True, preview only and never review or submit any
-            change, even when ``no_confirm`` is also set.
-        override: SHA hash to override non-automation change restriction.
-        topic: Explicit topic to scope the similar-change search to.
-            When omitted, the topic is taken from the search URL (if
-            given) or from the source change itself.
-        show_progress: If True, drive a live Rich progress tracker
-            during submission (in-place counters, matching the GitHub
-            merge path) instead of printing per-change status lines.
-    """
-    credentials = _resolve_gerrit_credentials_or_exit(
-        parsed_url, no_netrc, netrc_file, verbose, console
-    )
-
-    console.print(f"🔍 Examining Gerrit change on {parsed_url.host}...")
-
-    try:
-        service = create_gerrit_service(
-            host=parsed_url.host,
-            base_path=parsed_url.base_path,
-            username=credentials.username,
-            password=credentials.password,
-        )
-
-        if not service.is_authenticated:
-            console.print("⚠️ Warning: Service created but may not be authenticated")
-
-        source_change, topic_changes = _resolve_gerrit_source_change(
-            service, parsed_url, topic, credentials, console
-        )
-
-        comparator = create_gerrit_comparator(similarity_threshold=similarity_threshold)
-        only_automation = _resolve_gerrit_only_automation(
-            source_change, comparator, override, console
-        )
-
-        source_change = _maybe_rebase_gerrit_change(
-            service, source_change, credentials, console
-        )
-
-        candidates = _resolve_gerrit_candidates(
-            service, source_change, parsed_url, topic, topic_changes, console
-        )
-        similar_changes = _find_and_print_similar_changes(
-            service, comparator, source_change, candidates, only_automation, console
-        )
-
-        # Prepare list of changes to submit (similar + source)
-        source_entry: tuple[GerritChangeInfo, GerritComparisonResult | None] = (
-            source_change,
-            None,
-        )
-        all_changes: list[tuple[GerritChangeInfo, GerritComparisonResult | None]] = [
-            *similar_changes,
-            source_entry,
-        ]
-
-        if not _preview_gerrit_submission(
-            source_change, all_changes, no_confirm, dry_run, console
-        ):
-            return
-
-        _run_gerrit_submission(
-            parsed_url, credentials, all_changes, show_progress, console
-        )
-
-    except typer.Exit:
-        # Re-raise typer.Exit without treating it as an error
-        raise
-    except GerritAuthError as e:
-        console.print(f"❌ Gerrit authentication failed: {e}")
-        console.print("   Check your GERRIT_USERNAME and GERRIT_PASSWORD")
-        raise typer.Exit(1) from None
-    except GerritRestError as e:
-        console.print(f"❌ Gerrit API error: {e}")
-        raise typer.Exit(1) from None
-    except Exception as e:
-        console.print(f"❌ Error during Gerrit merge operation: {e}")
-        if verbose:
-            import traceback
-
-            traceback.print_exc()
-        raise typer.Exit(1) from None
-
-
-@app.command()
-def merge(
-    pr_url: str = typer.Argument(
-        ...,
-        help="GitHub PR URL, repository URL, owner/org URL, or Gerrit change URL",
-    ),
-    no_confirm: bool = typer.Option(
-        False,
-        "--no-confirm",
-        help="Skip confirmation prompt and merge immediately",
-    ),
-    similarity_threshold: float = typer.Option(
-        0.8, "--threshold", help="Similarity threshold for matching PRs (0.0-1.0)"
-    ),
-    merge_method: str = typer.Option(
-        "merge", "--merge-method", help="Merge method: merge, squash, or rebase"
-    ),
-    token: str | None = typer.Option(
-        None, "--token", help="GitHub token (or set GITHUB_TOKEN env var)"
-    ),
-    override: str | None = typer.Option(
-        None,
-        "--override",
-        help=(
-            "SHA hash authorizing a single non-automation PR/change. "
-            "Prefer --include-human-prs; this remains supported"
-        ),
-    ),
-    no_fix: bool = typer.Option(
-        False,
-        "--no-fix",
-        help="Do not attempt to automatically fix out-of-date branches",
-    ),
-    fix_semantic_title: bool = typer.Option(
-        True,
-        "--fix-semantic-title/--no-fix-semantic-title",
-        help=(
-            "Repair an automation PR whose title differs from its single "
-            "commit's subject, which permanently fails a semantic pull "
-            "request check. The title is set to the commit subject, and "
-            "GitHub's edited event re-runs the check. Only applied when "
-            "that check is the sole failure and the difference is an "
-            "elided version fragment. Default: enabled."
-        ),
-    ),
-    rebase_local: bool = typer.Option(
-        True,
-        "--rebase-local/--no-rebase-local",
-        help=(
-            "When rebasing a behind PR, prefer a local ``git`` clone + "
-            "rebase + force-push-with-lease over the GitHub REST "
-            "``update-branch`` endpoint when the base branch requires "
-            "verified signatures or the PR is from pre-commit-ci[bot]. "
-            "The local path inherits ``~/.gitconfig`` so commits stay "
-            "signed; the REST path is faster but produces unsigned "
-            "commits that break verification. Default: enabled."
-        ),
-    ),
-    merge_timeout: float = typer.Option(
-        DEFAULT_MERGE_TIMEOUT,
-        "--merge-timeout",
-        help=(
-            "Timeout in seconds for async merge operations (rebase, "
-            "pre-commit.ci, recreate). Default: "
-            f"{DEFAULT_MERGE_TIMEOUT:.0f}"
-        ),
-    ),
-    max_wait: float = typer.Option(
-        DEFAULT_MAX_WAIT,
-        "--max-wait",
-        help=(
-            "Owner/user-wide runs only: global wall-clock ceiling in "
-            "seconds for the whole batch. The run returns once every PR "
-            "has merged or this elapses (anything still in flight keeps "
-            "auto-merge armed and is reported as pending). 0 = "
-            "fire-and-forget (arm auto-merge and return immediately, "
-            f"never block). Default: {DEFAULT_MAX_WAIT:.0f}"
-        ),
-    ),
-    show_progress: bool = typer.Option(
-        True, "--progress/--no-progress", help="Show real-time progress updates"
-    ),
-    debug_matching: bool = typer.Option(
-        False,
-        "--debug-matching",
-        help="Show detailed scoring information for PR matching",
-    ),
-    dismiss_copilot: bool = typer.Option(
-        False,
-        "--dismiss-copilot",
-        help="Automatically dismiss unresolved GitHub Copilot review comments",
-    ),
-    force: str = typer.Option(
-        "code-owners",
-        "--force",
-        help="Override level: 'none', 'code-owners', 'protection-rules', 'all' (default: code-owners)",
-    ),
-    verbose: bool = typer.Option(
-        False,
-        "--verbose",
-        "-v",
-        help="Enable verbose debug logging",
-    ),
-    no_netrc: bool = typer.Option(
-        False,
-        "--no-netrc",
-        help="Disable .netrc credential lookup for Gerrit authentication",
-    ),
-    netrc_file: Path | None = typer.Option(
-        None,
-        "--netrc-file",
-        help="Explicit path to .netrc file for Gerrit credentials",
-        exists=True,
-        file_okay=True,
-        dir_okay=False,
-        readable=True,
-        resolve_path=True,
-    ),
-    netrc_optional: bool = typer.Option(
-        True,
-        "--netrc-optional/--netrc-required",
-        help="Whether to fail if .netrc file is not found (default: optional)",
-    ),
-    submit_gerrit_changes: bool = typer.Option(
-        False,
-        "--submit-gerrit-changes",
-        help="Explicitly request Gerrit submission for GitHub2Gerrit PRs (already the default when neither --skip-gerrit-changes nor --ignore-github2gerrit is given)",
-    ),
-    skip_gerrit_changes: bool = typer.Option(
-        False,
-        "--skip-gerrit-changes",
-        help="Skip PRs that have GitHub2Gerrit comments instead of merging them",
-    ),
-    ignore_github2gerrit: bool = typer.Option(
-        False,
-        "--ignore-github2gerrit",
-        help="Ignore GitHub2Gerrit comments and merge PRs in GitHub as normal",
-    ),
-    include_human_prs: bool = typer.Option(
-        False,
-        "--include-human-prs",
-        help=(
-            "Authorize human-authored PRs. Required when the source PR is "
-            "human-authored, and includes human-authored PRs in the similar-PR "
-            "set (prompting for confirmation unless --no-confirm)"
-        ),
-    ),
-    topic: str | None = typer.Option(
-        None,
-        "--topic",
-        help=(
-            "Gerrit only: scope the similar-change search to this topic. "
-            "When omitted, the topic is extracted automatically from a "
-            "Gerrit topic search URL or from the source change itself."
-        ),
-    ),
-    dry_run: bool = typer.Option(
-        False,
-        "--dry-run",
-        help=(
-            "Analyze and preview only: never approve, merge, rebase, or "
-            "close anything. Skips the write-permission pre-flight so it "
-            "runs under a read-only token (e.g. in CI). Implies "
-            "preview-only and suppresses confirmation prompts."
-        ),
-    ),
-):
-    """
-    Bulk approve/merge pull requests or Gerrit changes.
-
-    Supports GitHub PRs, GitHub repository URLs, and Gerrit Code Review changes.
-
-    By default, runs in interactive mode showing what changes will apply,
-    then prompts to proceed with merge. Use --no-confirm to merge immediately.
-
-    For GitHub PRs (single PR URL), this command will:
-
-    1. Analyze the provided PR
-
-    2. Find similar PRs across the owner (organization or user account)
-
-    3. Approve and merge matching PRs
-
-    4. Automatically fix out-of-date branches (use --no-fix to disable)
-
-    For GitHub repository URLs, this command will:
-
-    1. Fetch all open PRs in the specified repository
-
-    2. Filter to automation PRs only (unless --include-human-prs is given)
-
-    3. Approve and merge matching PRs in bulk
-
-    Repository URL formats accepted:
-      https://github.com/owner/repo
-      https://github.com/owner/repo/
-      https://github.com/owner/repo/pulls
-
-    For GitHub owner (organization or user) URLs, this command will:
-
-    1. Enumerate every non-archived, non-fork repository owned by the
-       organization or user (forks are excluded)
-
-    2. Fetch their open automation PRs (unless --include-human-prs)
-
-    3. Bulk merge them with a striped scheduler that processes at most
-       one PR per repository at a time, spreading work across
-       repositories to avoid racing GitHub's mergeability propagation
-
-    Owner URL formats accepted:
-      https://github.com/owner
-      https://github.com/owner/
-      https://github.com/orgs/owner
-      https://github.com/orgs/owner/repositories
-
-    For Gerrit changes, this command will:
-
-    1. Analyze the provided change
-
-    2. Find similar open changes on the server (scoped to the change's
-       topic when it has one, or the --topic flag)
-
-    3. Review (+2 Code-Review) and submit matching changes
-
-    Gerrit topic search URLs are also accepted; all open changes
-    sharing the topic form the batch:
-      https://gerrit.example.org/q/topic:some-topic
-      https://gerrit.onap.org/r/q/topic:some-topic
-
-    Merges similar PRs/changes from the same automation tool.
-
-    For user generated bulk PRs, use the --override flag with SHA hash.
-
-    GitHub2Gerrit handling:
-    By default, PRs with GitHub2Gerrit mapping comments are detected and
-    the corresponding Gerrit changes are submitted (+2 Code-Review + submit).
-    Use --skip-gerrit-changes to skip these PRs, or --ignore-github2gerrit
-    to merge them in GitHub as normal (which may leave orphaned Gerrit changes).
-
-    GitHub Force levels:
-    - none: Respect all protections
-    - code-owners: Bypass code owner review requirements (default)
-    - protection-rules: Bypass branch protection checks (requires permissions)
-    - all: Attempt merge despite most warnings (not recommended)
-
-    Authentication (Gerrit):
-    Credentials are loaded in this order:
-    1. .netrc file (if not disabled with --no-netrc)
-    2. Environment variables: GERRIT_USERNAME and GERRIT_PASSWORD
-
-    .netrc search order: ./netrc, ~/.netrc, ~/_netrc (Windows)
-    Use --netrc-file to specify an explicit path.
-    """
-    github2gerrit_mode = _validate_merge_inputs(
-        submit_gerrit_changes,
-        skip_gerrit_changes,
-        ignore_github2gerrit,
-        force,
-        verbose,
-    )
-
-    # Reject negative --max-wait.  The documented contract is 0 =
-    # fire-and-forget and > 0 = wall-clock ceiling; a negative value has
-    # no defined meaning and would otherwise be silently coerced into a
-    # surprising "instant no-wait" run (max_wait <= 0).  Fail fast so the
-    # flag's behaviour stays aligned with its help text.  The isinstance
-    # guard tolerates direct Python calls to ``merge`` (e.g. in tests),
-    # where Typer's ``OptionInfo`` default object is passed unresolved.
-    if isinstance(max_wait, int | float) and max_wait < 0:
-        console.print(
-            "❌ Invalid --max-wait: must be 0 (fire-and-forget) or a "
-            "positive number of seconds"
-        )
-        raise typer.Exit(1)
-
-    # Tolerate direct Python calls to ``merge`` (e.g. in tests), where
-    # Typer's ``OptionInfo`` default object is passed unresolved and
-    # would otherwise be truthy.
-    if not isinstance(topic, str):
-        topic = None
-    elif not topic.strip():
-        topic = None
-    else:
-        topic = topic.strip()
-
-    # Try as a specific PR/change URL first, then a Gerrit topic search
-    # URL, then an owner-wide URL (bare owner / orgs/owner), then a
-    # single repository URL.
-    parsed_url: ParsedUrl | None = None
-    parsed_topic: ParsedGerritTopicUrl | None = None
-    parsed_repo: ParsedRepoUrl | None = None
-    parsed_org: ParsedOrgUrl | None = None
-    change_err: UrlParseError | None = None
-    try:
-        parsed_url = parse_change_url(pr_url)
-    except UrlParseError as e:
-        change_err = e
-        # Not a PR/change URL — try a Gerrit topic search URL next, so
-        # pasted dashboard URLs like /q/topic:some-topic work directly.
-        try:
-            parsed_topic = parse_gerrit_topic_url(pr_url)
-        except UrlParseError:
-            parsed_topic = None
-    if parsed_url is None and parsed_topic is None and change_err is not None:
-        # Not a PR URL — try owner-wide before repository.  parse_org_url
-        # is strict (only a bare owner or the canonical orgs/owner forms),
-        # so a two-segment owner/repo URL falls through to parse_repo_url.
-        # Trying owner-wide first is required so /orgs/owner is not
-        # mis-parsed by parse_repo_url as owner="orgs", repo="owner".
-        try:
-            parsed_org = parse_org_url(pr_url)
-        except UrlParseError as org_err:
-            # Not an owner URL — try as a repository URL
-            try:
-                parsed_repo = parse_repo_url(pr_url)
-            except UrlParseError as repo_err:
-                # Show the most relevant error: if the URL targets a
-                # non-github.com host the original parse_change_url error
-                # gives host-appropriate guidance (e.g. Gerrit tips),
-                # whereas parse_repo_url only talks about github.com.
-                from .url_parser import _host_matches
-
-                # Prepend scheme if missing so urlparse can extract the
-                # hostname.  Without a scheme, schemeless URLs like
-                # "gerrit.example.org/..." are parsed as a path with no
-                # hostname, causing the wrong error to be shown.
-                _norm = pr_url
-                if not _norm.startswith(("http://", "https://")):
-                    _norm = "https://" + _norm
-                try:
-                    host = urlparse(_norm).hostname or ""
-                except Exception:
-                    host = ""
-                if host and not _host_matches(host.lower(), "github.com"):
-                    # Non-github host.  An owner-shaped path (``/orgs/owner``
-                    # or a single bare segment) most likely means the user
-                    # aimed an owner-wide URL at a non-github host (e.g.
-                    # GHE), so surface parse_org_url's actionable rejection
-                    # ("Owner-wide URL parsing is only supported for
-                    # github.com … use a direct PR URL") instead of the
-                    # generic parse_change_url "cannot determine platform"
-                    # message.  Any other shape (including Gerrit-style
-                    # URLs) keeps the platform-agnostic guidance.
-                    segs = [s for s in urlparse(_norm).path.split("/") if s]
-                    if segs and (segs[0] == "orgs" or len(segs) == 1):
-                        console.print(f"❌ Invalid URL: {org_err}")
-                    else:
-                        console.print(f"❌ Invalid URL: {change_err}")
-                else:
-                    console.print(f"❌ Invalid URL: {repo_err}")
-                raise typer.Exit(1) from None
-
-    if topic and not (
-        (parsed_url is not None and parsed_url.is_gerrit) or parsed_topic is not None
-    ):
-        console.print("❌ --topic is only supported for Gerrit URLs")
-        raise typer.Exit(1)
-
-    if parsed_topic is not None or (parsed_url is not None and parsed_url.is_gerrit):
-        gerrit_target: ParsedUrl | ParsedGerritTopicUrl
-        if parsed_topic is not None:
-            gerrit_target = parsed_topic
-        else:
-            assert parsed_url is not None
-            gerrit_target = parsed_url
-        _handle_gerrit_merge(
-            parsed_url=gerrit_target,
-            no_confirm=no_confirm,
-            similarity_threshold=similarity_threshold,
-            verbose=verbose,
-            console=console,
-            no_netrc=no_netrc,
-            netrc_file=netrc_file,
-            netrc_optional=netrc_optional,
-            dry_run=dry_run,
-            override=override,
-            topic=topic,
-            show_progress=show_progress,
-        )
-        return
-
-    if parsed_org is not None:
-        org_ctx = _MergeContext(
-            pr_url=parsed_org.original_url,
-            max_wait=max_wait,
-            no_confirm=no_confirm,
-            similarity_threshold=similarity_threshold,
-            merge_method=merge_method,
-            token=token,
-            override=override,
-            no_fix=no_fix,
-            fix_semantic_title=fix_semantic_title,
-            merge_timeout=merge_timeout,
-            show_progress=show_progress,
-            debug_matching=debug_matching,
-            dismiss_copilot=dismiss_copilot,
-            force=force,
-            verbose=verbose,
-            no_netrc=no_netrc,
-            netrc_file=netrc_file,
-            netrc_optional=netrc_optional,
-            github2gerrit_mode=github2gerrit_mode,
-            include_human_prs=include_human_prs,
-            rebase_local=rebase_local,
-            dry_run=dry_run,
-        )
-        try:
-            _handle_org_merge(parsed_org, org_ctx)
-        except DependamergeError as exc:
-            if org_ctx.progress_tracker:
-                org_ctx.progress_tracker.stop()
-            exc.display_and_exit()
-        except (KeyboardInterrupt, SystemExit):
-            if org_ctx.progress_tracker:
-                org_ctx.progress_tracker.stop()
-            raise
-        except typer.Exit:
-            if org_ctx.progress_tracker:
-                org_ctx.progress_tracker.stop()
-            raise
-        except (
-            GitError,
-            RateLimitError,
-            SecondaryRateLimitError,
-            GraphQLError,
-        ) as exc:
-            if org_ctx.progress_tracker:
-                org_ctx.progress_tracker.stop()
-            if isinstance(exc, GitError):
-                converted_error = convert_git_error(exc)
-            else:
-                converted_error = convert_github_api_error(exc)
-            converted_error.display_and_exit()
-        except Exception as e:
-            if org_ctx.progress_tracker:
-                org_ctx.progress_tracker.stop()
-            if is_github_api_permission_error(e):
-                exit_for_github_api_error(exception=e)
-            elif is_network_error(e):
-                converted_error = convert_network_error(e)
-                converted_error.display_and_exit()
-            else:
-                exit_with_error(
-                    ExitCode.GENERAL_ERROR,
-                    message="❌ Error during owner-wide merge operation",
-                    details=str(e),
-                    exception=e,
-                )
-        return
-
-    if parsed_repo is not None:
-        repo_ctx = _MergeContext(
-            pr_url=parsed_repo.original_url,
-            max_wait=max_wait,
-            no_confirm=no_confirm,
-            similarity_threshold=similarity_threshold,
-            merge_method=merge_method,
-            token=token,
-            override=override,
-            no_fix=no_fix,
-            fix_semantic_title=fix_semantic_title,
-            merge_timeout=merge_timeout,
-            show_progress=show_progress,
-            debug_matching=debug_matching,
-            dismiss_copilot=dismiss_copilot,
-            force=force,
-            verbose=verbose,
-            no_netrc=no_netrc,
-            netrc_file=netrc_file,
-            netrc_optional=netrc_optional,
-            github2gerrit_mode=github2gerrit_mode,
-            include_human_prs=include_human_prs,
-            rebase_local=rebase_local,
-            dry_run=dry_run,
-        )
-        try:
-            _handle_repo_merge(parsed_repo, repo_ctx)
-        except DependamergeError as exc:
-            if repo_ctx.progress_tracker:
-                repo_ctx.progress_tracker.stop()
-            exc.display_and_exit()
-        except (KeyboardInterrupt, SystemExit):
-            if repo_ctx.progress_tracker:
-                repo_ctx.progress_tracker.stop()
-            raise
-        except typer.Exit:
-            if repo_ctx.progress_tracker:
-                repo_ctx.progress_tracker.stop()
-            raise
-        except (
-            GitError,
-            RateLimitError,
-            SecondaryRateLimitError,
-            GraphQLError,
-        ) as exc:
-            if repo_ctx.progress_tracker:
-                repo_ctx.progress_tracker.stop()
-            if isinstance(exc, GitError):
-                converted_error = convert_git_error(exc)
-            else:
-                converted_error = convert_github_api_error(exc)
-            converted_error.display_and_exit()
-        except Exception as e:
-            if repo_ctx.progress_tracker:
-                repo_ctx.progress_tracker.stop()
-            if is_github_api_permission_error(e):
-                exit_for_github_api_error(exception=e)
-            elif is_network_error(e):
-                converted_error = convert_network_error(e)
-                converted_error.display_and_exit()
-            else:
-                exit_with_error(
-                    ExitCode.GENERAL_ERROR,
-                    message="❌ Error during repository merge operation",
-                    details=str(e),
-                    exception=e,
-                )
-        return
-
-    assert parsed_url is not None
-    ctx = _MergeContext(
-        pr_url=parsed_url.original_url,
-        max_wait=max_wait,
-        no_confirm=no_confirm,
-        similarity_threshold=similarity_threshold,
-        merge_method=merge_method,
-        token=token,
-        override=override,
-        no_fix=no_fix,
-        fix_semantic_title=fix_semantic_title,
-        merge_timeout=merge_timeout,
-        show_progress=show_progress,
-        debug_matching=debug_matching,
-        dismiss_copilot=dismiss_copilot,
-        force=force,
-        verbose=verbose,
-        no_netrc=no_netrc,
-        netrc_file=netrc_file,
-        netrc_optional=netrc_optional,
-        github2gerrit_mode=github2gerrit_mode,
-        include_human_prs=include_human_prs,
-        rebase_local=rebase_local,
-        dry_run=dry_run,
-    )
-
-    try:
-        _init_github_merge(ctx)
-        _fetch_and_validate_source_pr(ctx)
-        _maybe_check_merge_permissions(ctx)
-
-        # Debug matching info for source PR
-        if ctx.debug_matching:
-            _print_debug_matching(ctx)
-
-        _validate_automation_author(ctx)
-
-        # Scan org and find similar PRs
-        _scan_and_find_similar(ctx)
-
-        if not ctx.no_confirm or ctx.dry_run:
-            console.print("\n🔍 Dependamerge Evaluation\n")
-
-        assert ctx.source_pr is not None
-        source_entry: tuple[PullRequestInfo, ComparisonResult | None] = (
-            ctx.source_pr,
-            None,
-        )
-        all_prs_to_merge: list[tuple[PullRequestInfo, ComparisonResult | None]] = [
-            *ctx.all_similar_prs,
-            source_entry,
-        ]
-        # For the real merge (``--no-confirm``) the scan has already
-        # stopped the progress tracker; stand up a fresh one so the
-        # background wait-status ticker has a live display to update
-        # while PRs sit in the Step 5.5 auto-merge wait.  Preview and
-        # dry runs keep the stopped tracker (one line per PR, no wait
-        # loop) because they never execute a real merge.
-        if ctx.no_confirm and not ctx.dry_run:
-            _restart_merge_progress_tracker(ctx, len(all_prs_to_merge))
-        try:
-            merge_results = _run_parallel_merge(
-                ctx,
-                all_prs_to_merge,
-                preview=ctx.dry_run or not ctx.no_confirm,
-                # No similar-PR list was printed when none were found, so
-                # skip the blank line before the merge banner.
-                leading_blank=bool(ctx.all_similar_prs),
-            )
-        finally:
-            if (
-                ctx.no_confirm
-                and not ctx.dry_run
-                and ctx.show_progress
-                and ctx.progress_tracker
-            ):
-                ctx.progress_tracker.stop()
-
-        if not merge_results:
-            console.print("❌ No PRs were processed")
-            return
-
-        merged_count = sum(1 for r in merge_results if r.status.value == "merged")
-
-        # Dry run: report what *would* happen and stop before any prompt
-        # or real merge.  ``no_confirm=False`` selects the "Would …"
-        # preview phrasing in the results summary.
-        if ctx.dry_run:
-            _display_merge_results(merge_results, no_confirm=False)
-            return
-
-        if not ctx.no_confirm:
-            _handle_preview_confirmation(
-                ctx,
-                merge_results,
-                all_prs_to_merge,
-                merged_count,
-                len(merge_results),
-            )
-            return
-
-        _display_merge_results(merge_results, ctx.no_confirm)
-
-    except DependamergeError as exc:
-        if ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-        exc.display_and_exit()
-    except (KeyboardInterrupt, SystemExit):
-        if ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-        raise
-    except typer.Exit:
-        if ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-        raise
-    except (
-        GitError,
-        RateLimitError,
-        SecondaryRateLimitError,
-        GraphQLError,
-    ) as exc:
-        if ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-        if isinstance(exc, GitError):
-            converted_error = convert_git_error(exc)
-        else:
-            converted_error = convert_github_api_error(exc)
-        converted_error.display_and_exit()
-    except Exception as e:
-        if ctx.progress_tracker:
-            ctx.progress_tracker.stop()
-        if is_github_api_permission_error(e):
-            exit_for_github_api_error(exception=e)
-        elif is_network_error(e):
-            converted_error = convert_network_error(e)
-            converted_error.display_and_exit()
-        else:
-            exit_with_error(
-                ExitCode.GENERAL_ERROR,
-                message="❌ Error during merge operation",
-                details=str(e),
-                exception=e,
-            )
-
-
-def _print_debug_matching(ctx: _MergeContext) -> None:
-    """Print debug matching information for the source PR."""
-    assert ctx.github_client is not None
-    assert ctx.source_pr is not None
-    assert ctx.comparator is not None
-
-    console.print("\n🔍 Debug Matching Information")
-    console.print(
-        "   Source PR automation status: "
-        f"{ctx.github_client.is_automation_author(ctx.source_pr.author)}"
-    )
-    console.print(
-        "   Extracted package: "
-        f"'{ctx.comparator._extract_package_name(ctx.source_pr.title)}'"
-    )
-    console.print(f"   Similarity threshold: {ctx.similarity_threshold}")
-    if ctx.source_pr.body:
-        console.print(f"   Body preview: {ctx.source_pr.body[:100]}...")
-        console.print(
-            "   Is dependabot body: "
-            f"{ctx.comparator._is_dependabot_body(ctx.source_pr.body)}"
-        )
-    else:
-        console.print("   ⚠️ Source PR has no body")
-    console.print()
-
-
-def _display_pr_info(
-    pr: PullRequestInfo,
-    title: str,
-    github_client: GitHubClient,
-    progress_tracker: ProgressTracker | None = None,
-) -> None:
-    """Display pull request information in a formatted table."""
-    table = Table(title=title)
-    table.add_column("Property", style="cyan")
-    table.add_column("Value", style="green")
-
-    status = github_client.get_pr_status_details(pr)
-
-    table.add_row("Repository", pr.repository_full_name)
-    table.add_row("PR Number", str(pr.number))
-    table.add_row("Title", pr.title)
-    table.add_row("Author", pr.author)
-    table.add_row("State", pr.state)
-    table.add_row("Status", status)
-    table.add_row("Files Changed", str(len(pr.files_changed)))
-    table.add_row("URL", pr.html_url)
-
-    if progress_tracker:
-        progress_tracker.suspend()
-    console.print(table)
-    if progress_tracker:
-        progress_tracker.resume()
-
-
-@dataclass
-class _CloseContext:
-    """Shared state for the ``close`` command pipeline."""
-
-    token: str
-    github_client: GitHubClient
-    owner: str
-    repo_name: str
-    pr_number: int
-    source_pr: PullRequestInfo
-    progress_tracker: MergeProgressTracker | None
-
-
-def _run_close_parallel(
-    ctx: _CloseContext,
-    prs: list[PullRequestInfo],
-    preview_mode: bool,
-) -> list[CloseResult]:
-    """Close (or preview-close) ``prs`` in parallel."""
-
-    async def _close_parallel() -> list[CloseResult]:
-        close_manager = AsyncCloseManager(
-            token=ctx.token,
-            progress_tracker=ctx.progress_tracker,
-            preview_mode=preview_mode,
-        )
-        async with close_manager:
-            # Convert to list of tuples (PR, None) for consistency
-            pr_tuples: list[tuple[PullRequestInfo, ComparisonResult | None]] = [
-                (pr, None) for pr in prs
-            ]
-            return await close_manager.close_prs_parallel(pr_tuples)
-
-    return asyncio.run(_close_parallel())
-
-
-def _print_close_debug_matching(
-    ctx: _CloseContext,
-    comparator: PRComparator,
-    similarity_threshold: float,
-) -> None:
-    """Show detailed matching diagnostics for the source PR."""
-    source_pr = ctx.source_pr
-    console.print("\n🔍 Debug Matching Information")
-    console.print(
-        "   Source PR automation status: "
-        f"{ctx.github_client.is_automation_author(source_pr.author)}"
-    )
-    console.print(
-        f"   Extracted package: '{comparator._extract_package_name(source_pr.title)}'"
-    )
-    console.print(f"   Similarity threshold: {similarity_threshold}")
-    if source_pr.body:
-        console.print(f"   Body preview: {source_pr.body[:100]}...")
-        console.print(
-            f"   Is dependabot body: {comparator._is_dependabot_body(source_pr.body)}"
-        )
-    else:
-        console.print("   ⚠️ Source PR has no body")
-    console.print()
-
-
-def _validate_close_authorization(
-    ctx: _CloseContext,
-    override: str | None,
-) -> bool:
-    """Gate non-automation source PRs behind the override SHA.
-
-    Returns True when the close may proceed; False when the caller
-    should stop (the reason has already been printed).
-    """
-    if ctx.github_client.is_automation_author(ctx.source_pr.author):
-        return True
-
-    commit_messages = ctx.github_client.get_pull_request_commits(
-        ctx.owner, ctx.repo_name, ctx.pr_number
-    )
-    first_commit_line = commit_messages[0].split("\n")[0] if commit_messages else ""
-
-    # Generate expected SHA
-    expected_sha = _generate_override_sha(ctx.source_pr, first_commit_line)
-
-    if not override:
-        console.print("Source PR is not from a recognized automation tool.")
-        console.print(
-            f"To close this and similar PRs, run again with: --override {expected_sha}"
-        )
-        console.print(
-            f"This SHA is based on the author '{ctx.source_pr.author}' and commit message '{first_commit_line[:50]}...'",
-            style="dim",
-        )
-        return False
-
-    if override != expected_sha:
-        console.print(
-            f"Error: Invalid override SHA. Expected: {expected_sha}",
-            style="bold red",
-        )
-        console.print(
-            "This prevents accidental bulk operations on non-automation PRs.",
-            style="dim",
-        )
-        return False
-
-    console.print("Override SHA validated. Proceeding with non-automation PR close.")
-    return True
-
-
-def _find_similar_prs_for_close(
-    ctx: _CloseContext,
-    comparator: PRComparator,
-    debug_matching: bool,
-) -> list[tuple[PullRequestInfo, ComparisonResult]]:
-    """Search the owner for PRs similar to the source PR."""
-    if ctx.progress_tracker:
-        console.print()
-    else:
-        console.print(f"\nChecking owner: {ctx.owner}")
-
-    # Use GitHubService for async PR finding
-    from .github_service import GitHubService
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.update_operation("Listing repositories...")
-
-    async def _find_similar():
-        svc = GitHubService(
-            token=ctx.token,
-            progress_tracker=ctx.progress_tracker,
-            debug_matching=debug_matching,
-        )
-        try:
-            only_automation = ctx.github_client.is_automation_author(
-                ctx.source_pr.author
-            )
-            return await svc.find_similar_prs(
-                ctx.owner,
-                ctx.source_pr,
-                comparator,
-                only_automation=only_automation,
-            )
-        finally:
-            await svc.close()
-
-    return asyncio.run(_find_similar())
-
-
-def _print_close_analysis_summary(
-    ctx: _CloseContext,
-    all_similar_prs: list[tuple[PullRequestInfo, ComparisonResult]],
-) -> None:
-    """Report the search outcome and list the matching PRs."""
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-        summary = ctx.progress_tracker.get_summary()
-        elapsed_time = summary.get("elapsed_time")
-        total_prs_analyzed = summary.get("total_prs_analyzed")
-        completed_repositories = summary.get("completed_repositories")
-        similar_prs_found = summary.get("similar_prs_found")
-        errors_count = summary.get("errors_count", 0)
-        console.print(f"\n✅ Analysis completed in {elapsed_time}")
-        console.print(
-            f"📊 Analyzed {total_prs_analyzed} PRs across {completed_repositories} repositories"
-        )
-        console.print(f"🔍 Found {similar_prs_found} similar PRs")
-        if errors_count > 0:
-            console.print(f"⚠️ {errors_count} errors encountered during analysis")
-        console.print()
-    else:
-        console.print(f"\n🔍 Found {len(all_similar_prs)} similar PRs")
-
-    if not all_similar_prs:
-        # Not a failure: the supplied PR is simply the only one to
-        # close.  Use a neutral glyph rather than ❌ so the output
-        # does not read like an error.
-        console.print("⏩ No similar PRs found for this owner")
-
-    for target_pr, comparison in all_similar_prs:
-        console.print(f"  • {target_pr.repository_full_name} #{target_pr.number}")
-        console.print(f"    {_format_condensed_similarity(comparison)}")
-
-
-def _run_close_dry_run(
-    ctx: _CloseContext,
-    all_prs_to_close: list[PullRequestInfo],
-) -> None:
-    """Evaluate in preview mode and report what *would* be closed."""
-    if ctx.progress_tracker:
-        ctx.progress_tracker.start()
-        console.print()
-    else:
-        console.print(
-            f"\n🧪 Dry run: evaluating {len(all_prs_to_close)} pull requests..."
-        )
-
-    close_results = _run_close_parallel(ctx, all_prs_to_close, True)
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-        console.print()
-
-    closeable_count = sum(1 for r in close_results if r.status.value == "closed")
-    console.print(
-        f"\n🧪 Dry run: would close {closeable_count}/"
-        f"{len(all_prs_to_close)} PRs (no changes made)"
-    )
-
-
-def _run_interactive_close(
-    ctx: _CloseContext,
-    all_prs_to_close: list[PullRequestInfo],
-) -> None:
-    """Preview the close, then prompt for SHA-gated confirmation."""
-    if ctx.progress_tracker:
-        ctx.progress_tracker.start()
-        console.print()
-    else:
-        console.print(f"\n🚀 Evaluating {len(all_prs_to_close)} pull requests...")
-
-    close_results = _run_close_parallel(ctx, all_prs_to_close, True)
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-        console.print()
-
-    # Count closeable PRs
-    closed_count = sum(1 for r in close_results if r.status.value == "closed")
-    total_to_close = len(all_prs_to_close)
-
-    console.print(f"\nCloseable {closed_count}/{total_to_close} PRs")
-
-    if closed_count == 0:
-        console.print("\n❌ No PRs are eligible for closing")
-        return
-
-    # Generate continuation SHA and prompt user
-    commit_messages = ctx.github_client.get_pull_request_commits(
-        ctx.owner, ctx.repo_name, ctx.pr_number
-    )
-    first_commit_line = commit_messages[0].split("\n")[0] if commit_messages else ""
-    continue_sha_hash = _generate_continue_sha(ctx.source_pr, first_commit_line)
-    console.print()
-    console.print(f"To proceed with closing enter: {continue_sha_hash}")
-
-    # Check if in test mode (don't prompt during tests)
-    if "pytest" in sys.modules or os.getenv("TESTING"):
-        console.print("⚠️ Test mode detected - skipping interactive prompt")
-        return
-
-    user_input = typer.prompt(
-        "\nEnter the string above to continue (or press Enter to cancel)"
-    ).strip()
-
-    if user_input != continue_sha_hash:
-        console.print("\n❌ Operation cancelled by user")
-        return
-
-    # Run actual close on closeable PRs only
-    console.print(f"\n🔨 Closing {closed_count} closeable pull requests...")
-    closeable_prs = [
-        all_prs_to_close[i]
-        for i, result in enumerate(close_results)
-        # These were preview "closed"
-        if result.status.value == "closed"
-    ]
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.start()
-
-    final_results = _run_close_parallel(ctx, closeable_prs, False)
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-
-    # Count final results
-    final_closed = sum(1 for r in final_results if r.status.value == "closed")
-    final_failed = sum(1 for r in final_results if r.status.value == "failed")
-
-    console.print(f"\n🚀 Final Results: {final_closed} closed, {final_failed} failed")
-
-
-def _run_immediate_close(
-    ctx: _CloseContext,
-    all_prs_to_close: list[PullRequestInfo],
-) -> None:
-    """Close all candidate PRs without confirmation."""
-    if ctx.progress_tracker:
-        ctx.progress_tracker.start()
-    console.print(f"\n🚀 Closing {len(all_prs_to_close)} pull requests...")
-
-    close_results = _run_close_parallel(ctx, all_prs_to_close, False)
-
-    if ctx.progress_tracker:
-        ctx.progress_tracker.stop()
-
-    # Count results
-    closed_count = sum(1 for r in close_results if r.status.value == "closed")
-    failed_count = sum(1 for r in close_results if r.status.value == "failed")
-
-    console.print(f"\n🚀 Final Results: {closed_count} closed, {failed_count} failed")
-
-
-@app.command()
-def close(
-    pr_url: str = typer.Argument(..., help="GitHub pull request URL"),
-    no_confirm: bool = typer.Option(
-        False,
-        "--no-confirm",
-        help="Skip confirmation prompt and close immediately without preview",
-    ),
-    similarity_threshold: float = typer.Option(
-        0.8, "--threshold", help="Similarity threshold for matching PRs (0.0-1.0)"
-    ),
-    token: str | None = typer.Option(
-        None, "--token", help="GitHub token (or set GITHUB_TOKEN env var)"
-    ),
-    override: str | None = typer.Option(
-        None, "--override", help="SHA hash to override non-automation PR restriction"
-    ),
-    show_progress: bool = typer.Option(
-        True, "--progress/--no-progress", help="Show real-time progress updates"
-    ),
-    debug_matching: bool = typer.Option(
-        False,
-        "--debug-matching",
-        help="Show detailed scoring information for PR matching",
-    ),
-    dry_run: bool = typer.Option(
-        False,
-        "--dry-run",
-        help=(
-            "Analyze and preview only: never close anything. Suppresses "
-            "the confirmation prompt so it runs unattended under a "
-            "read-only token (e.g. in CI)."
-        ),
-    ),
-):
-    """
-    Bulk close pull requests across a GitHub owner (organization or user
-    account).
-
-    By default, runs in interactive mode showing what changes will apply,
-    then prompts to proceed with closing. Use --no-confirm to close immediately.
-
-    This command will:
-
-    1. Analyze the provided PR
-
-    2. Find similar PRs across the owner (organization or user account)
-
-    3. Close matching PRs
-
-    Closes similar PRs from the same automation tool (dependabot, pre-commit.ci).
-
-    For user generated bulk PRs, use the --override flag with SHA hash.
-    """
-    progress_tracker = None
-
-    try:
-        github_client = GitHubClient(token)
-        # GitHubClient resolves None -> GITHUB_TOKEN env var (raises if missing)
-        assert github_client.token is not None
-        token = github_client.token
-        owner, repo_name, pr_number = github_client.parse_pr_url(pr_url)
-
-        if show_progress:
-            progress_tracker = MergeProgressTracker(owner, is_close_operation=True)
-            progress_tracker.start()
-            # Check if Rich display is available
-            if not progress_tracker.rich_available:
-                console.print(f"🔍 Examining source pull request in {owner}...")
-                console.print("Progress updates will be shown as simple text...")
-        else:
-            console.print(f"🔍 Examining source pull request in {owner}...")
-
-        source_pr = github_client.get_pull_request_info(owner, repo_name, pr_number)
-
-        # Display source PR info
-        _display_pr_info(
-            source_pr, "", github_client, progress_tracker=progress_tracker
-        )
-
-        comparator = PRComparator(similarity_threshold)
-
-        ctx = _CloseContext(
-            token=token,
-            github_client=github_client,
-            owner=owner,
-            repo_name=repo_name,
-            pr_number=pr_number,
-            source_pr=source_pr,
-            progress_tracker=progress_tracker,
-        )
-
-        # Debug matching info for source PR
-        if debug_matching:
-            _print_close_debug_matching(ctx, comparator, similarity_threshold)
-
-        # Check if source PR is from automation or has valid override
-        if not _validate_close_authorization(ctx, override):
-            return
-
-        # Find similar PRs across the owner (organization or user)
-        all_similar_prs = _find_similar_prs_for_close(ctx, comparator, debug_matching)
-        _print_close_analysis_summary(ctx, all_similar_prs)
-
-        if not no_confirm:
-            # IMPORTANT: Each PR must produce exactly ONE line of output in this section
-            console.print("\n🔍 Dependamerge Evaluation\n")
-
-        # Determine which PRs to close
-        all_prs_to_close = [source_pr] + [pr for pr, _ in all_similar_prs]
-
-        if dry_run:
-            # Dry run: evaluate in preview mode and report what *would*
-            # be closed, then stop.  No prompt, no actual close — safe to
-            # run unattended under a read-only token.
-            _run_close_dry_run(ctx, all_prs_to_close)
-            return
-
-        if no_confirm:
-            # No confirmation - close immediately
-            _run_immediate_close(ctx, all_prs_to_close)
-        else:
-            _run_interactive_close(ctx, all_prs_to_close)
-
-    except DependamergeError as exc:
-        # Our structured errors handle display and exit themselves
-        if progress_tracker:
-            progress_tracker.stop()
-        exc.display_and_exit()
-    except (KeyboardInterrupt, SystemExit):
-        # Don't catch system interrupts or exits
-        if progress_tracker:
-            progress_tracker.stop()
-        raise
-    except typer.Exit:
-        if progress_tracker:
-            progress_tracker.stop()
-        # Re-raise without additional error messages
-        raise
-    except (GitError, RateLimitError, SecondaryRateLimitError, GraphQLError) as exc:
-        # Convert known errors to centralized error handling
-        if progress_tracker:
-            progress_tracker.stop()
-        if isinstance(exc, GitError):
-            converted_error = convert_git_error(exc)
-        else:  # GitHub API errors
-            converted_error = convert_github_api_error(exc)
-        converted_error.display_and_exit()
-    except Exception as e:
-        # Ensure progress tracker is stopped even if an unexpected error occurs
-        if progress_tracker:
-            progress_tracker.stop()
-
-        # Try to categorize the error
-        if is_github_api_permission_error(e):
-            exit_for_github_api_error(exception=e)
-        elif is_network_error(e):
-            converted_error = convert_network_error(e)
-            converted_error.display_and_exit()
-        else:
-            exit_with_error(
-                ExitCode.GENERAL_ERROR,
-                message="❌ Error during close operation",
-                details=str(e),
-                exception=e,
-            )
-
-
-@app.command()
-def status(
-    org_input: str = typer.Argument(
-        ...,
-        help="GitHub owner (organization or user) name or URL (e.g., 'lfreleng-actions' or 'https://github.com/lfreleng-actions/')",
-    ),
-    token: str | None = typer.Option(
-        None, "--token", help="GitHub token (or set GITHUB_TOKEN env var)"
-    ),
-    output_format: str = typer.Option(
-        "table", "--format", help="Output format: table, json"
-    ),
-    show_progress: bool = typer.Option(
-        True, "--progress/--no-progress", help="Show real-time progress updates"
-    ),
-):
-    """
-    Reports repository statistics for tags, releases and pull requests.
-
-    This command will:
-    1. Scan all repositories owned by the organization or user
-    2. Gather tag and release information
-    3. Count open and merged pull requests
-    4. Identify PRs affecting actions or workflows
-
-    Automation tools supported: Dependabot, Renovate, pre-commit.ci,
-    GitHub Actions, GitHub Copilot, and any other [bot] account.
-    """
-    # Parse owner login from input (handles a bare login plus every
-    # GitHub owner URL form, including /orgs/owner/repositories).
-    try:
-        org_name = parse_owner_arg(org_input)
-    except UrlParseError:
-        org_name = ""
-    if not org_name:
-        console.print("❌ Invalid GitHub owner name or URL")
-        console.print(
-            "   Expected an organization or user account, e.g. "
-            "'owner-name' or 'https://github.com/owner-name/'"
-        )
-        raise typer.Exit(1)
-
-    # Initialize progress tracker (disable PR stats for status command)
-    progress_tracker = None
-
-    try:
-        if show_progress:
-            progress_tracker = ProgressTracker(org_name, show_pr_stats=False)
-            progress_tracker.start()
-            if not progress_tracker.rich_available:
-                console.print(f"🔍 Scanning owner: {org_name}")
-                console.print("Progress updates will be shown as simple text...")
-        else:
-            console.print(f"🔍 Scanning owner: {org_name}")
-            console.print(
-                "This may take a few minutes for owners with many repositories..."
-            )
-
-        # Perform the scan
-        from .github_service import GitHubService
-
-        async def _run_status_check():
-            svc = GitHubService(token=token, progress_tracker=progress_tracker)
-            try:
-                return await svc.gather_organization_status(org_name)
-            finally:
-                await svc.close()
-
-        status_result = asyncio.run(_run_status_check())
-
-        if progress_tracker:
-            progress_tracker.stop()
-            if progress_tracker.rich_available:
-                console.print()
-            else:
-                console.print()
-
-            # Show scan summary
-            summary = progress_tracker.get_summary()
-            elapsed_time = summary.get("elapsed_time")
-            console.print(f"\n✅ Scan completed in {elapsed_time}")
-            console.print()
-
-        # Display results
-        _display_status_results(status_result, output_format)
-
-    except KeyboardInterrupt:
-        if progress_tracker:
-            progress_tracker.stop()
-        console.print("\n⚠️ Scan interrupted by user")
-        raise typer.Exit(130) from None
-    except Exception as e:
-        if progress_tracker:
-            progress_tracker.stop()
-        console.print(f"❌ Error during scan: {e}")
-        raise typer.Exit(1) from e
-
-
-@app.command()
-def blocked(
-    org_input: str = typer.Argument(
-        ...,
-        help="GitHub owner (organization or user) name or URL (e.g., 'lfreleng-actions' or 'https://github.com/lfreleng-actions/')",
-    ),
-    token: str | None = typer.Option(
-        None, "--token", help="GitHub token (or set GITHUB_TOKEN env var)"
-    ),
-    output_format: str = typer.Option(
-        "table", "--format", help="Output format: table, json"
-    ),
-    include_drafts: bool = typer.Option(
-        False,
-        "--include-drafts",
-        help="Include draft pull requests in the blocked PRs report",
-    ),
-    fix: bool = typer.Option(
-        False,
-        "--fix",
-        help="Interactively rebase to resolve conflicts and force-push updates",
-    ),
-    limit: int | None = typer.Option(
-        None, "--limit", help="Maximum number of PRs to attempt fixing"
-    ),
-    reason: str | None = typer.Option(
-        None,
-        "--reason",
-        help="Only fix PRs with this blocking reason (e.g., merge_conflict, behind_base)",
-    ),
-    workdir: str | None = typer.Option(
-        None,
-        "--workdir",
-        help="Base directory for workspaces (defaults to a secure temp dir)",
-    ),
-    keep_temp: bool = typer.Option(
-        False,
-        "--keep-temp",
-        help="Keep the temporary workspace for inspection after completion",
-    ),
-    prefetch: int | None = typer.Option(
-        None,
-        "--prefetch",
-        help="Number of repositories to prepare in parallel (auto-detects CPU cores if not specified)",
-    ),
-    editor: str | None = typer.Option(
-        None,
-        "--editor",
-        help="Editor command to use for resolving conflicts (defaults to $VISUAL or $EDITOR)",
-    ),
-    mergetool: bool = typer.Option(
-        False,
-        "--mergetool",
-        help="Use 'git mergetool' for resolving conflicts when available",
-    ),
-    interactive: bool = typer.Option(
-        True,
-        "--interactive/--no-interactive",
-        help="Attach rebase to the terminal for interactive resolution",
-    ),
-    show_progress: bool = typer.Option(
-        True, "--progress/--no-progress", help="Show real-time progress updates"
-    ),
-):
-    """
-    Reports blocked pull requests in a GitHub organization or user account.
-
-    This command will:
-    1. Check all repositories owned by the organization or user
-    2. Identify pull requests that cannot be merged
-    3. Report blocking reasons (conflicts, failing checks, etc.)
-    4. Count unresolved Copilot feedback comments
-
-    Standard code review requirements are not considered blocking.
-    """
-    # Parse owner login from input (handles a bare login plus every
-    # GitHub owner URL form, including /orgs/owner/repositories).
-    try:
-        organization = parse_owner_arg(org_input)
-    except UrlParseError:
-        organization = ""
-    if not organization:
-        console.print("❌ Invalid GitHub owner name or URL")
-        console.print(
-            "   Expected an organization or user account, e.g. "
-            "'owner-name' or 'https://github.com/owner-name/'"
-        )
-        raise typer.Exit(1)
-
-    progress_tracker = None
-
-    try:
-        if show_progress:
-            progress_tracker = ProgressTracker(organization)
-            progress_tracker.start()
-            # Check if Rich display is available
-            if not progress_tracker.rich_available:
-                console.print(f"🔍 Checking owner: {organization}")
-                console.print("Progress updates will be shown as simple text...")
-        else:
-            console.print(f"🔍 Checking owner: {organization}")
-            console.print(
-                "This may take a few minutes for owners with many repositories..."
-            )
-
-        # Perform the scan
-        from .github_service import GitHubService
-
-        async def _run_blocked_check():
-            svc = GitHubService(token=token, progress_tracker=progress_tracker)
-            try:
-                return await svc.scan_organization(
-                    organization, include_drafts=include_drafts
-                )
-            finally:
-                await svc.close()
-
-        scan_result = asyncio.run(_run_blocked_check())
-
-        if progress_tracker:
-            progress_tracker.stop()
-            if progress_tracker.rich_available:
-                console.print()  # Add blank line after progress display
-            else:
-                console.print()  # Clear the fallback display line
-
-            # Show scan summary
-            summary = progress_tracker.get_summary()
-            elapsed_time = summary.get("elapsed_time")
-            total_prs_analyzed = summary.get("total_prs_analyzed")
-            completed_repositories = summary.get("completed_repositories")
-            errors_count = summary.get("errors_count", 0)
-            console.print(f"✅ Check completed in {elapsed_time}")
-            console.print(
-                f"📊 Analyzed {total_prs_analyzed} PRs across {completed_repositories} repositories"
-            )
-            if errors_count > 0:
-                console.print(f"⚠️ {errors_count} errors encountered during check")
-            console.print()  # Add blank line before results
-
-        # Display results
-        _display_blocked_results(scan_result, output_format)
-
-        # Optional fix workflow
-        if fix:
-            allowed_default = {"merge_conflict", "behind_base"}
-            reasons_to_attempt = (
-                allowed_default if not reason else {reason.strip().lower()}
-            )
-
-            selections: list[PRSelection] = []
-            for pr in scan_result.unmergeable_prs:
-                pr_reason_types = {r.type for r in pr.reasons}
-                if pr_reason_types & reasons_to_attempt:
-                    selections.append(
-                        PRSelection(repository=pr.repository, pr_number=pr.pr_number)
-                    )
-
-            if limit is not None and limit > 0:
-                selections = selections[:limit]
-
-            if not selections:
-                console.print("No eligible PRs to fix based on the selected reasons.")
-                return
-
-            token_to_use = token or os.getenv("GITHUB_TOKEN")
-            if not token_to_use:
-                exit_for_configuration_error(
-                    message="❌ GitHub token required for --fix option",
-                    details="Provide --token or set GITHUB_TOKEN environment variable",
-                )
-
-            console.print(f"Starting interactive fix for {len(selections)} PR(s)...")
-            try:
-                orchestrator = FixOrchestrator(
-                    token_to_use,
-                    progress_tracker=progress_tracker,
-                    logger=lambda m: console.print(m),
-                )
-                fix_options = FixOptions(
-                    workdir=workdir,
-                    keep_temp=keep_temp,
-                    prefetch=prefetch
-                    if prefetch is not None
-                    else get_default_workers(),
-                    editor=editor,
-                    mergetool=mergetool,
-                    interactive=interactive,
-                    logger=lambda m: console.print(m),
-                )
-                results = orchestrator.run(selections, fix_options)
-                success_count = sum(1 for r in results if r.success)
-                console.print(
-                    f"✅ Fix complete: {success_count}/{len(selections)} succeeded"
-                )
-            except Exception as e:
-                exit_with_error(
-                    ExitCode.GENERAL_ERROR,
-                    message="❌ Error during fix workflow",
-                    details=str(e),
-                    exception=e,
-                )
-
-    except DependamergeError as exc:
-        # Our structured errors handle display and exit themselves
-        if progress_tracker:
-            progress_tracker.stop()
-        exc.display_and_exit()
-    except (KeyboardInterrupt, SystemExit):
-        # Don't catch system interrupts or exits
-        if progress_tracker:
-            progress_tracker.stop()
-        raise
-    except typer.Exit as e:
-        if progress_tracker:
-            progress_tracker.stop()
-        raise e
-    except (GitError, RateLimitError, SecondaryRateLimitError, GraphQLError) as exc:
-        # Convert known errors to centralized error handling
-        if progress_tracker:
-            progress_tracker.stop()
-        if isinstance(exc, GitError):
-            converted_error = convert_git_error(exc)
-        else:  # GitHub API errors
-            converted_error = convert_github_api_error(exc)
-        converted_error.display_and_exit()
-    except Exception as e:
-        # Ensure progress tracker is stopped even if an error occurs
-        if progress_tracker:
-            progress_tracker.stop()
-
-        # Try to categorize the error
-        if is_github_api_permission_error(e):
-            exit_for_github_api_error(exception=e)
-        elif is_network_error(e):
-            converted_error = convert_network_error(e)
-            converted_error.display_and_exit()
-        else:
-            exit_with_error(
-                ExitCode.GENERAL_ERROR,
-                message="❌ Error during owner scan",
-                details=str(e),
-                exception=e,
-            )
-
-
-def _display_blocked_results(scan_result, output_format: str):
-    """Display the organization blocked PR results."""
-
-    if output_format == "json":
-        import json
-
-        console.print(json.dumps(scan_result.model_dump(), indent=2, default=str))
-        return
-
-    # Table format
-    if not scan_result.unmergeable_prs:
-        console.print("🎉 No unmergeable pull requests found!")
-        return
-
-    pr_table = Table(title=f"Blocked Pull Requests: {scan_result.organization}")
-    pr_table.add_column("Repository", style="cyan")
-    pr_table.add_column("PR", style="white")
-    pr_table.add_column("Title", style="white", max_width=40)
-    pr_table.add_column("Author", style="white")
-    pr_table.add_column("Blocking Reasons", style="yellow")
-
-    # Only show Copilot column if there are any copilot comments
-    show_copilot_col = any(
-        p.copilot_comments_count > 0 for p in scan_result.unmergeable_prs
-    )
-    if show_copilot_col:
-        pr_table.add_column("Copilot", style="blue")
-
-    for pr in scan_result.unmergeable_prs:
-        reasons = [reason.description for reason in pr.reasons]
-        reasons_text = "\n".join(reasons) if reasons else "Unknown"
-
-        row_data = [
-            pr.repository.split("/", 1)[1] if "/" in pr.repository else pr.repository,
-            f"#{pr.pr_number}",
-            pr.title,
-            pr.author,
-            reasons_text,
-        ]
-
-        # Add Copilot count if column is shown
-        if show_copilot_col:
-            row_data.append(str(pr.copilot_comments_count))
-
-        pr_table.add_row(*row_data)
-
-    console.print(pr_table)
-    console.print()
-
-    # Create summary table (moved to bottom)
-    summary_table = Table()
-    summary_table.add_column("Summary", style="cyan")
-    summary_table.add_column("Value", style="white")
-
-    summary_table.add_row("Total Repositories", str(scan_result.total_repositories))
-    summary_table.add_row("Checked Repositories", str(scan_result.scanned_repositories))
-    summary_table.add_row("Total Open PRs", str(scan_result.total_prs))
-    summary_table.add_row("Unmergeable PRs", str(len(scan_result.unmergeable_prs)))
-
-    if scan_result.errors:
-        summary_table.add_row("Errors", str(len(scan_result.errors)), style="red")
-
-    console.print(summary_table)
-
-    # Show errors if any
-    if scan_result.errors:
-        console.print()
-        error_table = Table(title="Errors Encountered During Check")
-        error_table.add_column("Error", style="red")
-
-        for error in scan_result.errors:
-            error_table.add_row(error)
-
-        console.print(error_table)
-
-
-def _display_status_results(status_result, output_format: str):
-    """Display the organization status results."""
-
-    if output_format == "json":
-        import json
-
-        console.print(json.dumps(status_result.model_dump(), indent=2, default=str))
-        return
-
-    # Table format
-    if not status_result.repository_statuses:
-        console.print("❌ No repositories found in organization!")
-        return
-
-    status_table = Table(title=f"Organization: {status_result.organization}")
-    status_table.add_column("Repository", style="cyan")
-    status_table.add_column("Tag", style="white")
-    status_table.add_column("Date", style="white")
-    status_table.add_column("PRs Open", style="white")
-    status_table.add_column("PRs Merged", style="white")
-    status_table.add_column("Action", style="white")
-    status_table.add_column("Workflows", style="white")
-
-    for repo in status_result.repository_statuses:
-        # Format tag with icon
-        tag_display = "—"
-        if repo.latest_tag:
-            tag_display = f"{repo.status_icon} {repo.latest_tag}"
-
-        # Format date
-        date_display = repo.tag_date or repo.release_date or "—"
-
-        # Format PR counts
-        open_prs = f"{repo.open_prs_human} / {repo.open_prs_automation}"
-        merged_prs = f"{repo.merged_prs_human} / {repo.merged_prs_automation}"
-        action_prs = f"{repo.action_prs_human} / {repo.action_prs_automation}"
-        workflow_prs = f"{repo.workflow_prs_human} / {repo.workflow_prs_automation}"
-
-        status_table.add_row(
-            repo.repository_name,
-            tag_display,
-            date_display,
-            open_prs,
-            merged_prs,
-            action_prs,
-            workflow_prs,
-        )
-
-    console.print(status_table)
-    console.print()
-    console.print("PR counts are for human/automation")
-    console.print("\nAutomation tools supported:")
-    special_tool_labels = {
-        "[bot]": "Any other [bot] account",
-        "pre-commit": "pre-commit.ci",
-        "github-actions": "GitHub Actions",
-        "copilot": "GitHub Copilot",
-    }
-    for tool in AUTOMATION_TOOLS:
-        label = special_tool_labels.get(tool, tool.capitalize())
-        console.print(f"  • {label}")
-    console.print()
-
-    summary_table = Table()
-    summary_table.add_column("Summary", style="cyan")
-    summary_table.add_column("Value", style="white")
-
-    # Aggregate open-PR counts across all scanned repositories.  The
-    # per-repository "PRs Open" column shows human / automation, so the
-    # summary totals those same open-PR figures and reports the split
-    # before the combined total.
-    total_automation_prs = sum(
-        repo.open_prs_automation for repo in status_result.repository_statuses
-    )
-    total_human_prs = sum(
-        repo.open_prs_human for repo in status_result.repository_statuses
-    )
-
-    summary_table.add_row("🤖 Automation PRs", str(total_automation_prs))
-    summary_table.add_row("🤷 Human      PRs", str(total_human_prs))
-    summary_table.add_section()
-    summary_table.add_row("Total PRs", str(total_automation_prs + total_human_prs))
-    summary_table.add_row("Total Repositories", str(status_result.total_repositories))
-
-    # Only show Scanned Repositories if it differs from Total
-    if status_result.scanned_repositories != status_result.total_repositories:
-        summary_table.add_row(
-            "Scanned Repositories", str(status_result.scanned_repositories)
-        )
-
-    if status_result.errors:
-        summary_table.add_row("Errors", str(len(status_result.errors)), style="red")
-
-    console.print(summary_table)
-
-    # Show errors if any
-    if status_result.errors:
-        console.print()
-        error_table = Table(title="Errors Encountered During Scan")
-        error_table.add_column("Error", style="red")
-
-        for error in status_result.errors:
-            error_table.add_row(error)
-
-        console.print(error_table)
-
-
-if __name__ == "__main__":
-    app()
