+++ tests/test_fix_verification.py
@@ -15,9 +15,9 @@ from dependamerge.models import PullRequestInfo
 def test_final_verification():
     """Final test to demonstrate the fix is working"""
     with (
-        patch("dependamerge.cli._check_merge_permissions"),
-        patch("dependamerge.cli.GitHubClient") as mock_client_class,
-        patch("dependamerge.cli.PRComparator") as mock_comparator_class,
+        patch("dependamerge.cli._permissions._check_merge_permissions"),
+        patch("dependamerge.cli._deps.GitHubClient") as mock_client_class,
+        patch("dependamerge.cli._deps.PRComparator") as mock_comparator_class,
         patch("dependamerge.github_service.GitHubService") as mock_service_class,
         patch("dependamerge.merge_manager.GitHubAsync") as mock_async_class,
     ):
